decompose dashboard into components
