+++ _includes/portal/js/templates/dashboard/chart-js.jst
@@ -1,0 +1,56 @@
+<div class="ibox float-e-margins">
+  <div class="ibox-content">
+    <div>
+      <span class="pull-right text-right">
+        <small>Average value of sales in the past month in: <strong>United states</strong></small>
+        <br/>
+        All sales: 162,862
+      </span>
+      <h3 class="font-bold no-margins">
+        Half-year revenue margin
+      </h3>
+      <small>Sales marketing.</small>
+    </div>
+
+    <div class="m-t-sm">
+
+      <div class="row">
+        <div class="col-md-8">
+          <div>
+            <canvas id="lineChart" height="114"></canvas>
+          </div>
+        </div>
+        <div class="col-md-4">
+          <ul class="stat-list m-t-lg">
+            <li>
+              <h2 class="no-margins">2,346</h2>
+              <small>Total orders in period</small>
+              <div class="progress progress-mini">
+                <div class="progress-bar" style="width: 48%;"></div>
+              </div>
+            </li>
+            <li>
+              <h2 class="no-margins ">4,422</h2>
+              <small>Orders in last month</small>
+              <div class="progress progress-mini">
+                <div class="progress-bar" style="width: 60%;"></div>
+              </div>
+            </li>
+          </ul>
+        </div>
+      </div>
+
+    </div>
+
+    <div class="m-t-md">
+      <small class="pull-right">
+        <i class="fa fa-clock-o"> </i>
+        Update on 16.07.2015
+      </small>
+      <small>
+        <strong>Analysis of sales:</strong> The value has been changed over time, and last month reached a level over $50,000.
+      </small>
+    </div>
+
+  </div>
+</div>
