+++ _includes/portal/js/templates/dashboard/flot-chart.jst
@@ -1,0 +1,14 @@
+<div class="ibox float-e-margins">
+  <div class="ibox-title">
+    <h5>Monthly income</h5>
+    <div class="ibox-tools">
+      <span class="label label-primary">Updated 12.2015</span>
+    </div>
+  </div>
+  <div class="ibox-content no-padding">
+    <div class="flot-chart m-t-lg" style="height: 55px;">
+      <div class="flot-chart-content" id="flot-chart1"></div>
+    </div>
+  </div>
+
+</div>
