+++ _includes/portal/js/views/dashboard.js
@@ -1,135 +1,16 @@
 var Backbone = require('backbone');
+var TableView = require('./dashboard/table');
+var FlotChartView = require('./dashboard/flot-chart');
+var ChartJSView = require('./dashboard/chart-js');
 
 module.exports = Backbone.View.extend({
 
   template: require('../templates/dashboard.jst'),
 
   onShow: function() {
-    this.showFlotChart();
-    this.showChartJS();
-    this.showPieCharts();
-  },
-
-  showFlotChart: function() {
-    // flot chart
-    var d1 = [[1262304000000, 6], [1264982400000, 3057], [1267401600000, 20434], [1270080000000, 31982], [1272672000000, 26602], [1275350400000, 27826], [1277942400000, 24302], [1280620800000, 24237], [1283299200000, 21004], [1285891200000, 12144], [1288569600000, 10577], [1291161600000, 10295]];
-    var d2 = [[1262304000000, 5], [1264982400000, 200], [1267401600000, 1605], [1270080000000, 6129], [1272672000000, 11643], [1275350400000, 19055], [1277942400000, 30062], [1280620800000, 39197], [1283299200000, 37000], [1285891200000, 27000], [1288569600000, 21000], [1291161600000, 17000]];
-
-    var data1 = [
-        { label: 'Data 1', data: d1, color: '#17a084'},
-        { label: 'Data 2', data: d2, color: '#127e68' }
-    ];
-    if ($('#flot-chart1').length > 0) {
-      $.plot($('#flot-chart1'), data1, {
-          xaxis: {
-              tickDecimals: 0
-          },
-          series: {
-              lines: {
-                  show: true,
-                  fill: true,
-                  fillColor: {
-                      colors: [{
-                          opacity: 1
-                      }, {
-                          opacity: 1
-                      }]
-                  },
-              },
-              points: {
-                  width: 0.1,
-                  show: false
-              },
-          },
-          grid: {
-              show: false,
-              borderWidth: 0
-          },
-          legend: {
-              show: false,
-          }
-      });
-    }
-  },
-
-  showChartJS: function() {
-    var lineData = {
-      labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
-      datasets: [
-          {
-              label: 'Example dataset',
-              fillColor: 'rgba(220,220,220,0.5)',
-              strokeColor: 'rgba(220,220,220,1)',
-              pointColor: 'rgba(220,220,220,1)',
-              pointStrokeColor: '#fff',
-              pointHighlightFill: '#fff',
-              pointHighlightStroke: 'rgba(220,220,220,1)',
-              data: [65, 59, 40, 51, 36, 25, 40]
-          },
-          {
-              label: 'Example dataset',
-              fillColor: 'rgba(26,179,148,0.5)',
-              strokeColor: 'rgba(26,179,148,0.7)',
-              pointColor: 'rgba(26,179,148,1)',
-              pointStrokeColor: '#fff',
-              pointHighlightFill: '#fff',
-              pointHighlightStroke: 'rgba(26,179,148,1)',
-              data: [48, 48, 60, 39, 56, 37, 30]
-          }
-      ]
-    };
-
-    var lineOptions = {
-        scaleShowGridLines: true,
-        scaleGridLineColor: 'rgba(0,0,0,.05)',
-        scaleGridLineWidth: 1,
-        bezierCurve: true,
-        bezierCurveTension: 0.4,
-        pointDot: true,
-        pointDotRadius: 4,
-        pointDotStrokeWidth: 1,
-        pointHitDetectionRadius: 20,
-        datasetStroke: true,
-        datasetStrokeWidth: 2,
-        datasetFill: true,
-        responsive: true,
-    };
-
-    if (this.$('#lineChart').length > 0) {
-      var ctx = this.$('#lineChart')[0].getContext('2d');
-      var myNewChart = new Chart(ctx).Line(lineData, lineOptions);
-    }
-  },
-
-  showPieCharts: function() {
-    this.$('span.pie').peity('pie', {
-        fill: ['#1ab394', '#d7d7d7', '#ffffff']
-    });
-
-    this.$('.line').peity('line',{
-        fill: '#1ab394',
-        stroke:'#169c81',
-    });
-
-    this.$('.bar').peity('bar', {
-        fill: ['#1ab394', '#d7d7d7']
-    });
-
-    this.$('.bar_dashboard').peity('bar', {
-        fill: ['#1ab394', '#d7d7d7'],
-        width:100
-    });
-
-    var updatingChart = this.$('.updating-chart').peity('line', { fill: '#1ab394',stroke:'#169c81', width: 64 });
-
-    var random = Math.round(Math.random() * 10);
-    var values = updatingChart.text().split(',');
-    values.shift();
-    values.push(random);
-
-    updatingChart.text(values.join(',')).change();
-
+    this.addSubView({view: new TableView, selector: '#table'})
+    this.addSubView({view: new FlotChartView, selector: '#flot-chart'})
+    this.addSubView({view: new ChartJSView, selector: '#chart-js'})
   }
-
 
 });
