+++ js/app.js
@@ -1,5 +1,5 @@
-!function n(t,r,e){function i(s,l){if(!r[s]){if(!t[s]){var o="function"==typeof require&&require;if(!l&&o)return o(s,!0);if(a)return a(s,!0);var c=new Error("Cannot find module '"+s+"'");throw c.code="MODULE_NOT_FOUND",c}var u=r[s]={exports:{}};t[s][0].call(u.exports,function(n){var r=t[s][1][n];return i(r?r:n)},u,u.exports,n,t,r,e)}return r[s].exports}for(var a="function"==typeof require&&require,s=0;s<e.length;s++)i(e[s]);return i}({1:[function(n){var t=n("backbone"),r=(n("backbone.composer"),n("backbone.broker")),e=n("./router"),i=n("./views/main");t.$=window.$,t.View.prototype.attachToTemplate=!0,$(document).ready(function(){new i({el:"body"}),new e;r.start(),t.history.start()})},{"./router":3,"./views/main":15,backbone:22,"backbone.broker":18,"backbone.composer":20}],2:[function(n,t){var r=n("backbone");t.exports=r.Model.extend({defaults:{name:"David Williams"}})},{backbone:22}],3:[function(n,t){var r=n("backbone"),e=n("backbone.broker"),i=n("./views/dashboard"),a=n("./views/file-upload"),s=n("./views/file-manager"),l=e.channel("container"),o=e.channel("nav");t.exports=r.Router.extend({routes:{"":function(){l.publish("show",new i),o.publish("select","dashboard")},"file-manager":function(){l.publish("show",new s),o.publish("select","file-manager")},"file-upload":function(){l.publish("show",new a),o.publish("select","file-manager")}}})},{"./views/dashboard":11,"./views/file-manager":12,"./views/file-upload":13,backbone:22,"backbone.broker":18}],4:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>Dashboard</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content">\r\n  <div class="row">\r\n    <div class="col-md-2">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-success pull-right">Monthly</span>\r\n          <h5>Views</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <h1 class="no-margins">386,200</h1>\r\n          <div class="stat-percent font-bold text-success">98% <i class="fa fa-bolt"></i></div>\r\n          <small>Total views</small>\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-md-2">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-info pull-right">Annual</span>\r\n          <h5>Orders</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <h1 class="no-margins">80,800</h1>\r\n          <div class="stat-percent font-bold text-info">20% <i class="fa fa-level-up"></i></div>\r\n          <small>New orders</small>\r\n        </div>\r\n      </div>\r\n    </div>\r\n\r\n    <div class="col-md-4">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-primary pull-right">Today</span>\r\n          <h5>Vistits</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n\r\n          <div class="row">\r\n            <div class="col-md-6">\r\n              <h1 class="no-margins">$ 406,420</h1>\r\n              <div class="font-bold text-navy">44% <i class="fa fa-level-up"></i> <small>Rapid pace</small></div>\r\n            </div>\r\n            <div class="col-md-6">\r\n              <h1 class="no-margins">206,120</h1>\r\n              <div class="font-bold text-navy">22% <i class="fa fa-level-up"></i> <small>Slow pace</small></div>\r\n            </div>\r\n          </div>\r\n\r\n\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-md-4">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <h5>Monthly income</h5>\r\n          <div class="ibox-tools">\r\n            <span class="label label-primary">Updated 12.2015</span>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content no-padding">\r\n          <div class="flot-chart m-t-lg" style="height: 55px;">\r\n            <div class="flot-chart-content" id="flot-chart1"></div>\r\n          </div>\r\n        </div>\r\n\r\n      </div>\r\n    </div>\r\n  </div>\r\n  <div class="row">\r\n    <div class="col-lg-8">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-content">\r\n          <div>\r\n            <span class="pull-right text-right">\r\n              <small>Average value of sales in the past month in: <strong>United states</strong></small>\r\n              <br/>\r\n              All sales: 162,862\r\n            </span>\r\n            <h3 class="font-bold no-margins">\r\n              Half-year revenue margin\r\n            </h3>\r\n            <small>Sales marketing.</small>\r\n          </div>\r\n\r\n          <div class="m-t-sm">\r\n\r\n            <div class="row">\r\n              <div class="col-md-8">\r\n                <div>\r\n                  <canvas id="lineChart" height="114"></canvas>\r\n                </div>\r\n              </div>\r\n              <div class="col-md-4">\r\n                <ul class="stat-list m-t-lg">\r\n                  <li>\r\n                    <h2 class="no-margins">2,346</h2>\r\n                    <small>Total orders in period</small>\r\n                    <div class="progress progress-mini">\r\n                      <div class="progress-bar" style="width: 48%;"></div>\r\n                    </div>\r\n                  </li>\r\n                  <li>\r\n                    <h2 class="no-margins ">4,422</h2>\r\n                    <small>Orders in last month</small>\r\n                    <div class="progress progress-mini">\r\n                      <div class="progress-bar" style="width: 60%;"></div>\r\n                    </div>\r\n                  </li>\r\n                </ul>\r\n              </div>\r\n            </div>\r\n\r\n          </div>\r\n\r\n          <div class="m-t-md">\r\n            <small class="pull-right">\r\n              <i class="fa fa-clock-o"> </i>\r\n              Update on 16.07.2015\r\n            </small>\r\n            <small>\r\n              <strong>Analysis of sales:</strong> The value has been changed over time, and last month reached a level over $50,000.\r\n            </small>\r\n          </div>\r\n\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-lg-4">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-warning pull-right">Data has changed</span>\r\n          <h5>User activity</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>236 321.80</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>46.11%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>432.021</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>643 321.10</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>92.43%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>564.554</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>436 547.20</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>150.23%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>124.990</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n      </div>\r\n    </div>\r\n\r\n  </div>\r\n\r\n  <div class="row">\r\n\r\n    <div class="col-lg-12">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <h5>Custom responsive table </h5>\r\n          <div class="ibox-tools">\r\n            <a class="collapse-link">\r\n              <i class="fa fa-chevron-up"></i>\r\n            </a>\r\n            <a class="dropdown-toggle" data-toggle="dropdown" href="#">\r\n              <i class="fa fa-wrench"></i>\r\n            </a>\r\n            <ul class="dropdown-menu dropdown-user">\r\n              <li><a href="#">Config option 1</a>\r\n              </li>\r\n              <li><a href="#">Config option 2</a>\r\n              </li>\r\n            </ul>\r\n            <a class="close-link">\r\n              <i class="fa fa-times"></i>\r\n            </a>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-sm-9 m-b-xs">\r\n              <div data-toggle="buttons" class="btn-group">\r\n                <label class="btn btn-sm btn-white"> <input type="radio" id="option1" name="options"> Day </label>\r\n                <label class="btn btn-sm btn-white active"> <input type="radio" id="option2" name="options"> Week </label>\r\n                <label class="btn btn-sm btn-white"> <input type="radio" id="option3" name="options"> Month </label>\r\n              </div>\r\n            </div>\r\n            <div class="col-sm-3">\r\n              <div class="input-group"><input type="text" placeholder="Search" class="input-sm form-control"> <span class="input-group-btn">\r\n                <button type="button" class="btn btn-sm btn-primary"> Go!</button> </span></div>\r\n              </div>\r\n            </div>\r\n            <div class="table-responsive">\r\n              <table class="table table-striped">\r\n                <thead>\r\n                  <tr>\r\n\r\n                    <th>#</th>\r\n                    <th>Project </th>\r\n                    <th>Name </th>\r\n                    <th>Phone </th>\r\n                    <th>Company </th>\r\n                    <th>Completed </th>\r\n                    <th>Task</th>\r\n                    <th>Date</th>\r\n                    <th>Action</th>\r\n                  </tr>\r\n                </thead>\r\n                <tbody>\r\n                  <tr>\r\n                    <td>1</td>\r\n                    <td>Project <small>This is example of project</small></td>\r\n                    <td>Patrick Smith</td>\r\n                    <td>0800 051213</td>\r\n                    <td>Inceptos Hymenaeos Ltd</td>\r\n                    <td><span class="pie">0.52/1.561</span></td>\r\n                    <td>20%</td>\r\n                    <td>Jul 14, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>2</td>\r\n                    <td>Alpha project</td>\r\n                    <td>Alice Jackson</td>\r\n                    <td>0500 780909</td>\r\n                    <td>Nec Euismod In Company</td>\r\n                    <td><span class="pie">6,9</span></td>\r\n                    <td>40%</td>\r\n                    <td>Jul 16, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>3</td>\r\n                    <td>Betha project</td>\r\n                    <td>John Smith</td>\r\n                    <td>0800 1111</td>\r\n                    <td>Erat Volutpat</td>\r\n                    <td><span class="pie">3,1</span></td>\r\n                    <td>75%</td>\r\n                    <td>Jul 18, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>4</td>\r\n                    <td>Gamma project</td>\r\n                    <td>Anna Jordan</td>\r\n                    <td>(016977) 0648</td>\r\n                    <td>Tellus Ltd</td>\r\n                    <td><span class="pie">4,9</span></td>\r\n                    <td>18%</td>\r\n                    <td>Jul 22, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>2</td>\r\n                    <td>Alpha project</td>\r\n                    <td>Alice Jackson</td>\r\n                    <td>0500 780909</td>\r\n                    <td>Nec Euismod In Company</td>\r\n                    <td><span class="pie">6,9</span></td>\r\n                    <td>40%</td>\r\n                    <td>Jul 16, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>1</td>\r\n                    <td>Project <small>This is example of project</small></td>\r\n                    <td>Patrick Smith</td>\r\n                    <td>0800 051213</td>\r\n                    <td>Inceptos Hymenaeos Ltd</td>\r\n                    <td><span class="pie">0.52/1.561</span></td>\r\n                    <td>20%</td>\r\n                    <td>Jul 14, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>4</td>\r\n                    <td>Gamma project</td>\r\n                    <td>Anna Jordan</td>\r\n                    <td>(016977) 0648</td>\r\n                    <td>Tellus Ltd</td>\r\n                    <td><span class="pie">4,9</span></td>\r\n                    <td>18%</td>\r\n                    <td>Jul 22, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>1</td>\r\n                    <td>Project <small>This is example of project</small></td>\r\n                    <td>Patrick Smith</td>\r\n                    <td>0800 051213</td>\r\n                    <td>Inceptos Hymenaeos Ltd</td>\r\n                    <td><span class="pie">0.52/1.561</span></td>\r\n                    <td>20%</td>\r\n                    <td>Jul 14, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>2</td>\r\n                    <td>Alpha project</td>\r\n                    <td>Alice Jackson</td>\r\n                    <td>0500 780909</td>\r\n                    <td>Nec Euismod In Company</td>\r\n                    <td><span class="pie">6,9</span></td>\r\n                    <td>40%</td>\r\n                    <td>Jul 16, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>3</td>\r\n                    <td>Betha project</td>\r\n                    <td>John Smith</td>\r\n                    <td>0800 1111</td>\r\n                    <td>Erat Volutpat</td>\r\n                    <td><span class="pie">3,1</span></td>\r\n                    <td>75%</td>\r\n                    <td>Jul 18, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>4</td>\r\n                    <td>Gamma project</td>\r\n                    <td>Anna Jordan</td>\r\n                    <td>(016977) 0648</td>\r\n                    <td>Tellus Ltd</td>\r\n                    <td><span class="pie">4,9</span></td>\r\n                    <td>18%</td>\r\n                    <td>Jul 22, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>2</td>\r\n                    <td>Alpha project</td>\r\n                    <td>Alice Jackson</td>\r\n                    <td>0500 780909</td>\r\n                    <td>Nec Euismod In Company</td>\r\n                    <td><span class="pie">6,9</span></td>\r\n                    <td>40%</td>\r\n                    <td>Jul 16, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>1</td>\r\n                    <td>Project <small>This is example of project</small></td>\r\n                    <td>Patrick Smith</td>\r\n                    <td>0800 051213</td>\r\n                    <td>Inceptos Hymenaeos Ltd</td>\r\n                    <td><span class="pie">0.52/1.561</span></td>\r\n                    <td>20%</td>\r\n                    <td>Jul 14, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                  <tr>\r\n                    <td>4</td>\r\n                    <td>Gamma project</td>\r\n                    <td>Anna Jordan</td>\r\n                    <td>(016977) 0648</td>\r\n                    <td>Tellus Ltd</td>\r\n                    <td><span class="pie">4,9</span></td>\r\n                    <td>18%</td>\r\n                    <td>Jul 22, 2013</td>\r\n                    <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n                  </tr>\r\n                </tbody>\r\n              </table>\r\n            </div>\r\n\r\n          </div>\r\n        </div>\r\n      </div>\r\n    </div>\r\n  </div>\r\n';return __p}},{underscore:24}],5:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>Documents</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content">\r\n  <div class="row">\r\n    <div class="col-lg-3">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-content">\r\n          <div class="file-manager">\r\n            <h5>Show:</h5>\r\n            <a href="#" class="file-control active">Ale</a>\r\n            <a href="#" class="file-control">Documents</a>\r\n            <a href="#" class="file-control">Audio</a>\r\n            <a href="#" class="file-control">Images</a>\r\n            <div class="hr-line-dashed"></div>\r\n            <a href="#file-upload">\r\n              <button class="btn btn-primary btn-block">Upload Files</button>\r\n            </a>\r\n            <div class="hr-line-dashed"></div>\r\n            <h5>Folders</h5>\r\n            <ul class="folder-list" style="padding: 0">\r\n              <li><a href=""><i class="fa fa-folder"></i> Files</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Pictures</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Web pages</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Illustrations</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Films</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Books</a></li>\r\n            </ul>\r\n            <h5 class="tag-title">Tags</h5>\r\n            <ul class="tag-list" style="padding: 0">\r\n              <li><a href="">Family</a></li>\r\n              <li><a href="">Work</a></li>\r\n              <li><a href="">Home</a></li>\r\n              <li><a href="">Children</a></li>\r\n              <li><a href="">Holidays</a></li>\r\n              <li><a href="">Music</a></li>\r\n              <li><a href="">Photography</a></li>\r\n              <li><a href="">Film</a></li>\r\n            </ul>\r\n            <div class="clearfix"></div>\r\n          </div>\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-lg-9 animated fadeInRight">\r\n      <div class="row">\r\n        <div class="col-lg-12">\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n\r\n        </div>\r\n      </div>\r\n    </div>\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:24}],6:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";
-Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>File Upload</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content animated fadeIn">\r\n  <div class="row">\r\n    <div class="col-lg-12">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <h5>Dropzone Area</h5>\r\n          <div class="ibox-tools">\r\n            <a class="collapse-link">\r\n              <i class="fa fa-chevron-up"></i>\r\n            </a>\r\n            <a class="dropdown-toggle" data-toggle="dropdown" href="#">\r\n              <i class="fa fa-wrench"></i>\r\n            </a>\r\n            <ul class="dropdown-menu dropdown-user">\r\n              <li><a href="#">Config option 1</a>\r\n              </li>\r\n              <li><a href="#">Config option 2</a>\r\n              </li>\r\n            </ul>\r\n            <a class="close-link">\r\n              <i class="fa fa-times"></i>\r\n            </a>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <form id="my-awesome-dropzone" class="dropzone" action="#">\r\n            <div class="dropzone-previews"></div>\r\n            <button type="submit" class="btn btn-primary pull-right">Submit this form!</button>\r\n          </form>\r\n        </div>\r\n      </div>\r\n    </div>\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:24}],7:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="footer">\r\n  <div class="pull-right">\r\n    10GB of <strong>250GB</strong> Free.\r\n  </div>\r\n  <div>\r\n    <strong>Copyright</strong> Example Company &copy; 2014-2015\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:24}],8:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __t,__p="";Array.prototype.join}with(obj||{})__p+='<nav class="navbar-default navbar-static-side" role="navigation">\n    <div class="sidebar-collapse">\n        <ul class="nav" id="side-menu">\n            <li class="nav-header">\n                <div class="dropdown profile-element">\n                    <span>\n                        <img alt="image" class="img-circle" src="/img/portal/profile_small.jpg" />\n                    </span>\n                    <a data-toggle="dropdown" class="dropdown-toggle" href="#">\n                        <span class="clear">\n                          <span class="block m-t-xs">\n                            <strong class="font-bold">'+(null==(__t=name)?"":__t)+'</strong><b class="caret"></b>\n                          </span>\n                        </span>\n                    </a>\n                    <ul class="dropdown-menu animated fadeInRight m-t-xs">\n                        <li><a href="profile.html">Profile</a></li>\n                        <li><a href="contacts.html">Contacts</a></li>\n                        <li><a href="mailbox.html">Mailbox</a></li>\n                        <li class="divider"></li>\n                        <li><a href="login.html">Logout</a></li>\n                    </ul>\n                </div>\n            </li>\n            <li id="nav-dashboard">\n                <a href="#"><i class="fa fa-bar-chart-o"></i> <span class="nav-label">Dashboard</span></a>\n            </li>\n            <li id="nav-file-manager">\n                <a href="#file-manager"><i class="fa fa-files-o"></i> <span class="nav-label">Documents</span></a>\n            </li>\n        </ul>\n\n    </div>\n</nav>\n';return __p}},{underscore:24}],9:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row border-bottom">\r\n  <nav class="navbar navbar-static-top" role="navigation" style="margin-bottom: 0">\r\n    <div class="navbar-header">\r\n      <a class="navbar-minimalize minimalize-styl-2 btn btn-primary " href="#"><i class="fa fa-bars"></i> </a>\r\n    </div>\r\n    <ul class="nav navbar-top-links navbar-right">\r\n      <li>\r\n        <span class="m-r-sm text-muted welcome-message">Welcome to INSPINIA+ Admin Theme.</span>\r\n      </li>\r\n      <li class="dropdown">\r\n        <a class="dropdown-toggle count-info" data-toggle="dropdown" href="#">\r\n          <i class="fa fa-envelope"></i>  <span class="label label-warning">16</span>\r\n        </a>\r\n        <ul class="dropdown-menu dropdown-messages">\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/a7.jpg">\r\n              </a>\r\n              <div class="media-body">\r\n                <small class="pull-right">46h ago</small>\r\n                <strong>Mike Loreipsum</strong> started following <strong>Monica Smith</strong>. <br>\r\n                <small class="text-muted">3 days ago at 7:58 pm - 10.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/a4.jpg">\r\n              </a>\r\n              <div class="media-body ">\r\n                <small class="pull-right text-navy">5h ago</small>\r\n                <strong>Chris Johnatan Overtunk</strong> started following <strong>Monica Smith</strong>. <br>\r\n                <small class="text-muted">Yesterday 1:21 pm - 11.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/profile.jpg">\r\n              </a>\r\n              <div class="media-body ">\r\n                <small class="pull-right">23h ago</small>\r\n                <strong>Monica Smith</strong> love <strong>Kim Smith</strong>. <br>\r\n                <small class="text-muted">2 days ago at 2:30 am - 11.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="text-center link-block">\r\n              <a href="mailbox.html">\r\n                <i class="fa fa-envelope"></i> <strong>Read All Messages</strong>\r\n              </a>\r\n            </div>\r\n          </li>\r\n        </ul>\r\n      </li>\r\n      <li class="dropdown">\r\n        <a class="dropdown-toggle count-info" data-toggle="dropdown" href="#">\r\n          <i class="fa fa-bell"></i>  <span class="label label-primary">8</span>\r\n        </a>\r\n        <ul class="dropdown-menu dropdown-alerts">\r\n          <li>\r\n            <a href="mailbox.html">\r\n              <div>\r\n                <i class="fa fa-envelope fa-fw"></i> You have 16 messages\r\n                <span class="pull-right text-muted small">4 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <a href="profile.html">\r\n              <div>\r\n                <i class="fa fa-twitter fa-fw"></i> 3 New Followers\r\n                <span class="pull-right text-muted small">12 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <a href="grid_options.html">\r\n              <div>\r\n                <i class="fa fa-upload fa-fw"></i> Server Rebooted\r\n                <span class="pull-right text-muted small">4 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="text-center link-block">\r\n              <a href="notifications.html">\r\n                <strong>See All Alerts</strong>\r\n                <i class="fa fa-angle-right"></i>\r\n              </a>\r\n            </div>\r\n          </li>\r\n        </ul>\r\n      </li>\r\n\r\n\r\n      <li>\r\n        <a href="/login">\r\n          <i class="fa fa-sign-out"></i> Log out\r\n        </a>\r\n      </li>\r\n    </ul>\r\n\r\n  </nav>\r\n</div>\r\n';return __p}},{underscore:24}],10:[function(n,t){{var r=n("backbone"),e=n("backbone.broker");n("./file-upload")}t.exports=r.View.extend({initialize:function(){this.listenTo(e.channel("container"),"show",this.showView)},showView:function(n){this.setView(n)}})},{"./file-upload":13,backbone:22,"backbone.broker":18}],11:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/dashboard.jst"),onShow:function(){this.showFlotChart(),this.showChartJS(),this.showPieCharts()},showFlotChart:function(){var n=[[1262304e6,6],[12649824e5,3057],[12674016e5,20434],[127008e7,31982],[1272672e6,26602],[12753504e5,27826],[12779424e5,24302],[12806208e5,24237],[12832992e5,21004],[12858912e5,12144],[12885696e5,10577],[12911616e5,10295]],t=[[1262304e6,5],[12649824e5,200],[12674016e5,1605],[127008e7,6129],[1272672e6,11643],[12753504e5,19055],[12779424e5,30062],[12806208e5,39197],[12832992e5,37e3],[12858912e5,27e3],[12885696e5,21e3],[12911616e5,17e3]],r=[{label:"Data 1",data:n,color:"#17a084"},{label:"Data 2",data:t,color:"#127e68"}];$("#flot-chart1").length>0&&$.plot($("#flot-chart1"),r,{xaxis:{tickDecimals:0},series:{lines:{show:!0,fill:!0,fillColor:{colors:[{opacity:1},{opacity:1}]}},points:{width:.1,show:!1}},grid:{show:!1,borderWidth:0},legend:{show:!1}})},showChartJS:function(){var n={labels:["January","February","March","April","May","June","July"],datasets:[{label:"Example dataset",fillColor:"rgba(220,220,220,0.5)",strokeColor:"rgba(220,220,220,1)",pointColor:"rgba(220,220,220,1)",pointStrokeColor:"#fff",pointHighlightFill:"#fff",pointHighlightStroke:"rgba(220,220,220,1)",data:[65,59,40,51,36,25,40]},{label:"Example dataset",fillColor:"rgba(26,179,148,0.5)",strokeColor:"rgba(26,179,148,0.7)",pointColor:"rgba(26,179,148,1)",pointStrokeColor:"#fff",pointHighlightFill:"#fff",pointHighlightStroke:"rgba(26,179,148,1)",data:[48,48,60,39,56,37,30]}]},t={scaleShowGridLines:!0,scaleGridLineColor:"rgba(0,0,0,.05)",scaleGridLineWidth:1,bezierCurve:!0,bezierCurveTension:.4,pointDot:!0,pointDotRadius:4,pointDotStrokeWidth:1,pointHitDetectionRadius:20,datasetStroke:!0,datasetStrokeWidth:2,datasetFill:!0,responsive:!0};if(this.$("#lineChart").length>0){var r=this.$("#lineChart")[0].getContext("2d");new Chart(r).Line(n,t)}},showPieCharts:function(){this.$("span.pie").peity("pie",{fill:["#1ab394","#d7d7d7","#ffffff"]}),this.$(".line").peity("line",{fill:"#1ab394",stroke:"#169c81"}),this.$(".bar").peity("bar",{fill:["#1ab394","#d7d7d7"]}),this.$(".bar_dashboard").peity("bar",{fill:["#1ab394","#d7d7d7"],width:100});var n=this.$(".updating-chart").peity("line",{fill:"#1ab394",stroke:"#169c81",width:64}),t=Math.round(10*Math.random()),r=n.text().split(",");r.shift(),r.push(t),n.text(r.join(",")).change()}})},{"../templates/dashboard.jst":4,backbone:22}],12:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/file-manager.jst")})},{"../templates/file-manager.jst":5,backbone:22}],13:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/file-upload.jst")})},{"../templates/file-upload.jst":6,backbone:22}],14:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/footer.jst")})},{"../templates/footer.jst":7,backbone:22}],15:[function(n,t){var r=n("backbone"),e=n("./nav"),i=n("./top-nav"),a=n("./container"),s=n("./footer"),l=n("../models/user");t.exports=r.View.extend({initialize:function(){this.addSubView({view:new e({model:new l}),selector:"#wrapper",location:"prepend"}),this.addSubView({view:new i,selector:"#page-wrapper"}),this.addSubView({view:new a,selector:"#page-wrapper"}),this.addSubView({view:new s,selector:"#page-wrapper"})}})},{"../models/user":2,"./container":10,"./footer":14,"./nav":16,"./top-nav":17,backbone:22}],16:[function(n,t){var r=n("backbone"),e=n("backbone.broker");t.exports=r.View.extend({template:n("../templates/nav.jst"),initialize:function(){$(window).bind("load resize",function(){$(this).width()<769?$("body").addClass("body-small"):$("body").removeClass("body-small")}),this.listenTo(e.channel("nav"),"select",this.highlightSelection),this.listenTo(e.channel("nav"),"toggle",this.toggleNav),this.listenTo(this.model,"change",this.render)},serializeData:function(){return this.model.toJSON()},onShow:function(){$("body").hasClass("fixed-sidebar")&&this.$(".sidebar-collapse").slimScroll({height:"100%",railOpacity:.9})},highlightSelection:function(n){this.$("li").removeClass("active"),this.$("#nav-"+n).addClass("active")},toggleNav:function(){$("body").toggleClass("mini-navbar");var n=function(){this.$("#side-menu").fadeIn(500)}.bind(this);!$("body").hasClass("mini-navbar")||$("body").hasClass("body-small")?(this.$("#side-menu").hide(),setTimeout(n,100)):$("body").hasClass("fixed-sidebar")?(this.$("#side-menu").hide(),setTimeout(n,300)):this.$("#side-menu").removeAttr("style")},onClose:function(){$(window).off("load resize")}})},{"../templates/nav.jst":8,backbone:22,"backbone.broker":18}],17:[function(n,t){var r=n("backbone"),e=n("backbone.broker");t.exports=r.View.extend({template:n("../templates/top-nav.jst"),events:{"click .navbar-minimalize":"toggleNavBar"},toggleNavBar:function(n){n.preventDefault(),e.channel("nav").publish("toggle")}})},{"../templates/top-nav.jst":9,backbone:22,"backbone.broker":18}],18:[function(n,t,r){!function(t){"function"==typeof define&&define.amd?define(["underscore","backbone","exports"],t):"object"==typeof r?t(n("underscore"),n("backbone"),r):Backbone.broker||(Backbone.broker={},t(_,Backbone,Backbone.broker))}(function(n,t,r){var e=!1,i=[],a=function(n){this.name=n,this.queue=[]};n.extend(a.prototype,t.Events,{publish:function(){return arguments.length>0&&this.queue.push(arguments),e&&(this.queue.forEach(function(n){this.trigger.apply(this,n)}.bind(this)),this.queue=[]),this},subscribe:function(){return this.on.apply(this,arguments),this},unsubscribe:function(){return this.off.apply(this,arguments),this}}),n.extend(r,{channel:function(n){if(n&&n.length>0){var t=i.filter(function(t){return t.name===n?t:void 0});if(t.length>0)return t[0];var r=new a(n);return i.push(r),r}},getChannels:function(){return i},start:function(){e=!0,i.forEach(function(n){n.publish()})},stop:function(){e=!1}})})},{backbone:22,underscore:19}],19:[function(n,t,r){(function(){var n=this,e=n._,i=Array.prototype,a=Object.prototype,s=Function.prototype,l=i.push,o=i.slice,c=i.concat,u=a.toString,d=a.hasOwnProperty,f=Array.isArray,h=Object.keys,p=s.bind,v=function(n){return n instanceof v?n:this instanceof v?void(this._wrapped=n):new v(n)};"undefined"!=typeof r?("undefined"!=typeof t&&t.exports&&(r=t.exports=v),r._=v):n._=v,v.VERSION="1.7.0";var m=function(n,t,r){if(void 0===t)return n;switch(null==r?3:r){case 1:return function(r){return n.call(t,r)};case 2:return function(r,e){return n.call(t,r,e)};case 3:return function(r,e,i){return n.call(t,r,e,i)};case 4:return function(r,e,i,a){return n.call(t,r,e,i,a)}}return function(){return n.apply(t,arguments)}};v.iteratee=function(n,t,r){return null==n?v.identity:v.isFunction(n)?m(n,t,r):v.isObject(n)?v.matches(n):v.property(n)},v.each=v.forEach=function(n,t,r){if(null==n)return n;t=m(t,r);var e,i=n.length;if(i===+i)for(e=0;i>e;e++)t(n[e],e,n);else{var a=v.keys(n);for(e=0,i=a.length;i>e;e++)t(n[a[e]],a[e],n)}return n},v.map=v.collect=function(n,t,r){if(null==n)return[];t=v.iteratee(t,r);for(var e,i=n.length!==+n.length&&v.keys(n),a=(i||n).length,s=Array(a),l=0;a>l;l++)e=i?i[l]:l,s[l]=t(n[e],e,n);return s};var g="Reduce of empty array with no initial value";v.reduce=v.foldl=v.inject=function(n,t,r,e){null==n&&(n=[]),t=m(t,e,4);var i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length,l=0;if(arguments.length<3){if(!s)throw new TypeError(g);r=n[a?a[l++]:l++]}for(;s>l;l++)i=a?a[l]:l,r=t(r,n[i],i,n);return r},v.reduceRight=v.foldr=function(n,t,r,e){null==n&&(n=[]),t=m(t,e,4);var i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;if(arguments.length<3){if(!s)throw new TypeError(g);r=n[a?a[--s]:--s]}for(;s--;)i=a?a[s]:s,r=t(r,n[i],i,n);return r},v.find=v.detect=function(n,t,r){var e;return t=v.iteratee(t,r),v.some(n,function(n,r,i){return t(n,r,i)?(e=n,!0):void 0}),e},v.filter=v.select=function(n,t,r){var e=[];return null==n?e:(t=v.iteratee(t,r),v.each(n,function(n,r,i){t(n,r,i)&&e.push(n)}),e)},v.reject=function(n,t,r){return v.filter(n,v.negate(v.iteratee(t)),r)},v.every=v.all=function(n,t,r){if(null==n)return!0;t=v.iteratee(t,r);var e,i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;for(e=0;s>e;e++)if(i=a?a[e]:e,!t(n[i],i,n))return!1;return!0},v.some=v.any=function(n,t,r){if(null==n)return!1;t=v.iteratee(t,r);var e,i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;for(e=0;s>e;e++)if(i=a?a[e]:e,t(n[i],i,n))return!0;return!1},v.contains=v.include=function(n,t){return null==n?!1:(n.length!==+n.length&&(n=v.values(n)),v.indexOf(n,t)>=0)},v.invoke=function(n,t){var r=o.call(arguments,2),e=v.isFunction(t);return v.map(n,function(n){return(e?t:n[t]).apply(n,r)})},v.pluck=function(n,t){return v.map(n,v.property(t))},v.where=function(n,t){return v.filter(n,v.matches(t))},v.findWhere=function(n,t){return v.find(n,v.matches(t))},v.max=function(n,t,r){var e,i,a=-1/0,s=-1/0;if(null==t&&null!=n){n=n.length===+n.length?n:v.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],e>a&&(a=e)}else t=v.iteratee(t,r),v.each(n,function(n,r,e){i=t(n,r,e),(i>s||i===-1/0&&a===-1/0)&&(a=n,s=i)});return a},v.min=function(n,t,r){var e,i,a=1/0,s=1/0;if(null==t&&null!=n){n=n.length===+n.length?n:v.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],a>e&&(a=e)}else t=v.iteratee(t,r),v.each(n,function(n,r,e){i=t(n,r,e),(s>i||1/0===i&&1/0===a)&&(a=n,s=i)});return a},v.shuffle=function(n){for(var t,r=n&&n.length===+n.length?n:v.values(n),e=r.length,i=Array(e),a=0;e>a;a++)t=v.random(0,a),t!==a&&(i[a]=i[t]),i[t]=r[a];return i},v.sample=function(n,t,r){return null==t||r?(n.length!==+n.length&&(n=v.values(n)),n[v.random(n.length-1)]):v.shuffle(n).slice(0,Math.max(0,t))},v.sortBy=function(n,t,r){return t=v.iteratee(t,r),v.pluck(v.map(n,function(n,r,e){return{value:n,index:r,criteria:t(n,r,e)}}).sort(function(n,t){var r=n.criteria,e=t.criteria;if(r!==e){if(r>e||void 0===r)return 1;if(e>r||void 0===e)return-1}return n.index-t.index}),"value")};var b=function(n){return function(t,r,e){var i={};return r=v.iteratee(r,e),v.each(t,function(e,a){var s=r(e,a,t);n(i,e,s)}),i}};v.groupBy=b(function(n,t,r){v.has(n,r)?n[r].push(t):n[r]=[t]}),v.indexBy=b(function(n,t,r){n[r]=t}),v.countBy=b(function(n,t,r){v.has(n,r)?n[r]++:n[r]=1}),v.sortedIndex=function(n,t,r,e){r=v.iteratee(r,e,1);for(var i=r(t),a=0,s=n.length;s>a;){var l=a+s>>>1;r(n[l])<i?a=l+1:s=l}return a},v.toArray=function(n){return n?v.isArray(n)?o.call(n):n.length===+n.length?v.map(n,v.identity):v.values(n):[]},v.size=function(n){return null==n?0:n.length===+n.length?n.length:v.keys(n).length},v.partition=function(n,t,r){t=v.iteratee(t,r);var e=[],i=[];return v.each(n,function(n,r,a){(t(n,r,a)?e:i).push(n)}),[e,i]},v.first=v.head=v.take=function(n,t,r){return null==n?void 0:null==t||r?n[0]:0>t?[]:o.call(n,0,t)},v.initial=function(n,t,r){return o.call(n,0,Math.max(0,n.length-(null==t||r?1:t)))},v.last=function(n,t,r){return null==n?void 0:null==t||r?n[n.length-1]:o.call(n,Math.max(n.length-t,0))},v.rest=v.tail=v.drop=function(n,t,r){return o.call(n,null==t||r?1:t)},v.compact=function(n){return v.filter(n,v.identity)};var y=function(n,t,r,e){if(t&&v.every(n,v.isArray))return c.apply(e,n);for(var i=0,a=n.length;a>i;i++){var s=n[i];v.isArray(s)||v.isArguments(s)?t?l.apply(e,s):y(s,t,r,e):r||e.push(s)}return e};v.flatten=function(n,t){return y(n,t,!1,[])},v.without=function(n){return v.difference(n,o.call(arguments,1))},v.uniq=v.unique=function(n,t,r,e){if(null==n)return[];v.isBoolean(t)||(e=r,r=t,t=!1),null!=r&&(r=v.iteratee(r,e));for(var i=[],a=[],s=0,l=n.length;l>s;s++){var o=n[s];if(t)s&&a===o||i.push(o),a=o;else if(r){var c=r(o,s,n);v.indexOf(a,c)<0&&(a.push(c),i.push(o))}else v.indexOf(i,o)<0&&i.push(o)}return i},v.union=function(){return v.uniq(y(arguments,!0,!0,[]))},v.intersection=function(n){if(null==n)return[];for(var t=[],r=arguments.length,e=0,i=n.length;i>e;e++){var a=n[e];if(!v.contains(t,a)){for(var s=1;r>s&&v.contains(arguments[s],a);s++);s===r&&t.push(a)}}return t},v.difference=function(n){var t=y(o.call(arguments,1),!0,!0,[]);return v.filter(n,function(n){return!v.contains(t,n)})},v.zip=function(n){if(null==n)return[];for(var t=v.max(arguments,"length").length,r=Array(t),e=0;t>e;e++)r[e]=v.pluck(arguments,e);return r},v.object=function(n,t){if(null==n)return{};for(var r={},e=0,i=n.length;i>e;e++)t?r[n[e]]=t[e]:r[n[e][0]]=n[e][1];return r},v.indexOf=function(n,t,r){if(null==n)return-1;var e=0,i=n.length;if(r){if("number"!=typeof r)return e=v.sortedIndex(n,t),n[e]===t?e:-1;e=0>r?Math.max(0,i+r):r}for(;i>e;e++)if(n[e]===t)return e;return-1},v.lastIndexOf=function(n,t,r){if(null==n)return-1;var e=n.length;for("number"==typeof r&&(e=0>r?e+r+1:Math.min(e,r+1));--e>=0;)if(n[e]===t)return e;return-1},v.range=function(n,t,r){arguments.length<=1&&(t=n||0,n=0),r=r||1;for(var e=Math.max(Math.ceil((t-n)/r),0),i=Array(e),a=0;e>a;a++,n+=r)i[a]=n;return i};var w=function(){};v.bind=function(n,t){var r,e;if(p&&n.bind===p)return p.apply(n,o.call(arguments,1));if(!v.isFunction(n))throw new TypeError("Bind must be called on a function");return r=o.call(arguments,2),e=function(){if(!(this instanceof e))return n.apply(t,r.concat(o.call(arguments)));w.prototype=n.prototype;var i=new w;w.prototype=null;var a=n.apply(i,r.concat(o.call(arguments)));return v.isObject(a)?a:i}},v.partial=function(n){var t=o.call(arguments,1);return function(){for(var r=0,e=t.slice(),i=0,a=e.length;a>i;i++)e[i]===v&&(e[i]=arguments[r++]);for(;r<arguments.length;)e.push(arguments[r++]);return n.apply(this,e)}},v.bindAll=function(n){var t,r,e=arguments.length;if(1>=e)throw new Error("bindAll must be passed function names");for(t=1;e>t;t++)r=arguments[t],n[r]=v.bind(n[r],n);return n},v.memoize=function(n,t){var r=function(e){var i=r.cache,a=t?t.apply(this,arguments):e;return v.has(i,a)||(i[a]=n.apply(this,arguments)),i[a]};return r.cache={},r},v.delay=function(n,t){var r=o.call(arguments,2);return setTimeout(function(){return n.apply(null,r)},t)},v.defer=function(n){return v.delay.apply(v,[n,1].concat(o.call(arguments,1)))},v.throttle=function(n,t,r){var e,i,a,s=null,l=0;r||(r={});var o=function(){l=r.leading===!1?0:v.now(),s=null,a=n.apply(e,i),s||(e=i=null)};return function(){var c=v.now();l||r.leading!==!1||(l=c);var u=t-(c-l);return e=this,i=arguments,0>=u||u>t?(clearTimeout(s),s=null,l=c,a=n.apply(e,i),s||(e=i=null)):s||r.trailing===!1||(s=setTimeout(o,u)),a}},v.debounce=function(n,t,r){var e,i,a,s,l,o=function(){var c=v.now()-s;t>c&&c>0?e=setTimeout(o,t-c):(e=null,r||(l=n.apply(a,i),e||(a=i=null)))};return function(){a=this,i=arguments,s=v.now();var c=r&&!e;return e||(e=setTimeout(o,t)),c&&(l=n.apply(a,i),a=i=null),l}},v.wrap=function(n,t){return v.partial(t,n)},v.negate=function(n){return function(){return!n.apply(this,arguments)}},v.compose=function(){var n=arguments,t=n.length-1;return function(){for(var r=t,e=n[t].apply(this,arguments);r--;)e=n[r].call(this,e);return e}},v.after=function(n,t){return function(){return--n<1?t.apply(this,arguments):void 0}},v.before=function(n,t){var r;return function(){return--n>0?r=t.apply(this,arguments):t=null,r}},v.once=v.partial(v.before,2),v.keys=function(n){if(!v.isObject(n))return[];if(h)return h(n);var t=[];for(var r in n)v.has(n,r)&&t.push(r);return t},v.values=function(n){for(var t=v.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=n[t[i]];return e},v.pairs=function(n){for(var t=v.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=[t[i],n[t[i]]];return e},v.invert=function(n){for(var t={},r=v.keys(n),e=0,i=r.length;i>e;e++)t[n[r[e]]]=r[e];return t},v.functions=v.methods=function(n){var t=[];for(var r in n)v.isFunction(n[r])&&t.push(r);return t.sort()},v.extend=function(n){if(!v.isObject(n))return n;for(var t,r,e=1,i=arguments.length;i>e;e++){t=arguments[e];for(r in t)d.call(t,r)&&(n[r]=t[r])}return n},v.pick=function(n,t,r){var e,i={};if(null==n)return i;if(v.isFunction(t)){t=m(t,r);for(e in n){var a=n[e];t(a,e,n)&&(i[e]=a)}}else{var s=c.apply([],o.call(arguments,1));n=new Object(n);for(var l=0,u=s.length;u>l;l++)e=s[l],e in n&&(i[e]=n[e])}return i},v.omit=function(n,t,r){if(v.isFunction(t))t=v.negate(t);else{var e=v.map(c.apply([],o.call(arguments,1)),String);t=function(n,t){return!v.contains(e,t)}}return v.pick(n,t,r)},v.defaults=function(n){if(!v.isObject(n))return n;for(var t=1,r=arguments.length;r>t;t++){var e=arguments[t];for(var i in e)void 0===n[i]&&(n[i]=e[i])}return n},v.clone=function(n){return v.isObject(n)?v.isArray(n)?n.slice():v.extend({},n):n},v.tap=function(n,t){return t(n),n};var x=function(n,t,r,e){if(n===t)return 0!==n||1/n===1/t;if(null==n||null==t)return n===t;n instanceof v&&(n=n._wrapped),t instanceof v&&(t=t._wrapped);var i=u.call(n);if(i!==u.call(t))return!1;switch(i){case"[object RegExp]":case"[object String]":return""+n==""+t;case"[object Number]":return+n!==+n?+t!==+t:0===+n?1/+n===1/t:+n===+t;case"[object Date]":case"[object Boolean]":return+n===+t}if("object"!=typeof n||"object"!=typeof t)return!1;for(var a=r.length;a--;)if(r[a]===n)return e[a]===t;var s=n.constructor,l=t.constructor;if(s!==l&&"constructor"in n&&"constructor"in t&&!(v.isFunction(s)&&s instanceof s&&v.isFunction(l)&&l instanceof l))return!1;r.push(n),e.push(t);var o,c;if("[object Array]"===i){if(o=n.length,c=o===t.length)for(;o--&&(c=x(n[o],t[o],r,e)););}else{var d,f=v.keys(n);if(o=f.length,c=v.keys(t).length===o)for(;o--&&(d=f[o],c=v.has(t,d)&&x(n[d],t[d],r,e)););}return r.pop(),e.pop(),c};v.isEqual=function(n,t){return x(n,t,[],[])},v.isEmpty=function(n){if(null==n)return!0;if(v.isArray(n)||v.isString(n)||v.isArguments(n))return 0===n.length;for(var t in n)if(v.has(n,t))return!1;return!0},v.isElement=function(n){return!(!n||1!==n.nodeType)},v.isArray=f||function(n){return"[object Array]"===u.call(n)},v.isObject=function(n){var t=typeof n;return"function"===t||"object"===t&&!!n},v.each(["Arguments","Function","String","Number","Date","RegExp"],function(n){v["is"+n]=function(t){return u.call(t)==="[object "+n+"]"}}),v.isArguments(arguments)||(v.isArguments=function(n){return v.has(n,"callee")}),"function"!=typeof/./&&(v.isFunction=function(n){return"function"==typeof n||!1}),v.isFinite=function(n){return isFinite(n)&&!isNaN(parseFloat(n))},v.isNaN=function(n){return v.isNumber(n)&&n!==+n},v.isBoolean=function(n){return n===!0||n===!1||"[object Boolean]"===u.call(n)},v.isNull=function(n){return null===n},v.isUndefined=function(n){return void 0===n},v.has=function(n,t){return null!=n&&d.call(n,t)},v.noConflict=function(){return n._=e,this},v.identity=function(n){return n},v.constant=function(n){return function(){return n}},v.noop=function(){},v.property=function(n){return function(t){return t[n]}},v.matches=function(n){var t=v.pairs(n),r=t.length;return function(n){if(null==n)return!r;n=new Object(n);for(var e=0;r>e;e++){var i=t[e],a=i[0];if(i[1]!==n[a]||!(a in n))return!1}return!0}},v.times=function(n,t,r){var e=Array(Math.max(0,n));t=m(t,r,1);for(var i=0;n>i;i++)e[i]=t(i);return e},v.random=function(n,t){return null==t&&(t=n,n=0),n+Math.floor(Math.random()*(t-n+1))},v.now=Date.now||function(){return(new Date).getTime()};var _={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#x27;","`":"&#x60;"},k=v.invert(_),j=function(n){var t=function(t){return n[t]},r="(?:"+v.keys(n).join("|")+")",e=RegExp(r),i=RegExp(r,"g");return function(n){return n=null==n?"":""+n,e.test(n)?n.replace(i,t):n}};v.escape=j(_),v.unescape=j(k),v.result=function(n,t){if(null==n)return void 0;var r=n[t];return v.isFunction(r)?n[t]():r};var A=0;v.uniqueId=function(n){var t=++A+"";return n?n+t:t},v.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g,escape:/<%-([\s\S]+?)%>/g};var S=/(.)^/,E={"'":"'","\\":"\\","\r":"r","\n":"n","\u2028":"u2028","\u2029":"u2029"},T=/\\|'|\r|\n|\u2028|\u2029/g,O=function(n){return"\\"+E[n]};v.template=function(n,t,r){!t&&r&&(t=r),t=v.defaults({},t,v.templateSettings);var e=RegExp([(t.escape||S).source,(t.interpolate||S).source,(t.evaluate||S).source].join("|")+"|$","g"),i=0,a="__p+='";n.replace(e,function(t,r,e,s,l){return a+=n.slice(i,l).replace(T,O),i=l+t.length,r?a+="'+\n((__t=("+r+"))==null?'':_.escape(__t))+\n'":e?a+="'+\n((__t=("+e+"))==null?'':__t)+\n'":s&&(a+="';\n"+s+"\n__p+='"),t}),a+="';\n",t.variable||(a="with(obj||{}){\n"+a+"}\n"),a="var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n"+a+"return __p;\n";try{var s=new Function(t.variable||"obj","_",a)}catch(l){throw l.source=a,l}var o=function(n){return s.call(this,n,v)},c=t.variable||"obj";return o.source="function("+c+"){\n"+a+"}",o},v.chain=function(n){var t=v(n);return t._chain=!0,t};var C=function(n){return this._chain?v(n).chain():n};v.mixin=function(n){v.each(v.functions(n),function(t){var r=v[t]=n[t];v.prototype[t]=function(){var n=[this._wrapped];return l.apply(n,arguments),C.call(this,r.apply(v,n))}})},v.mixin(v),v.each(["pop","push","reverse","shift","sort","splice","unshift"],function(n){var t=i[n];v.prototype[n]=function(){var r=this._wrapped;return t.apply(r,arguments),"shift"!==n&&"splice"!==n||0!==r.length||delete r[0],C.call(this,r)}}),v.each(["concat","join","slice"],function(n){var t=i[n];v.prototype[n]=function(){return C.call(this,t.apply(this._wrapped,arguments))}}),v.prototype.value=function(){return this._wrapped},"function"==typeof define&&define.amd&&define("underscore",[],function(){return v})}).call(this)},{}],20:[function(n,t,r){!function(t){"function"==typeof define&&define.amd?define(["underscore","backbone"],t):"object"==typeof r?t(n("underscore"),n("backbone")):t(_,Backbone)}(function(n,t){n.extend(t.View.prototype,{render:function(){if(this.template&&n.isFunction(this.template)){var t=n.isFunction(this.serializeData)?this.serializeData():this,r=$(this.template(t));this.attachToTemplate&&1===r.length?(this.$el.replaceWith(r),this.setElement(r)):this.$el.html(r)}return this.restoreSubViews(),this.onRender&&n.isFunction(this.onRender)&&this.onRender.apply(this,arguments),this.trigger("rendered",this),this},setView:function(n,t){return this.removeSubViews(),t&&t.emptyDOM&&this.$el.empty(),this.addSubView({view:n,selector:this.$el}),n},getView:function(){return this._subViews&&this._subViews.length>0?this._subViews[0]:void 0},addSubView:function(n){if(!n||!n.view)throw new Error("Missing required view option");return this._subViews?this._subViews.push(n):this._subViews=[n],this.listenTo(n.view,"closed",this._removeSubView),this._showSubView(n)},restoreSubViews:function(){n.each(this._subViews,this._showSubView,this)},_showSubView:function(t){var r;return r=n.isObject(t.selector)?t.selector:n.isString(t.selector)?this.$(t.selector):this.$el,t.view.render(),"prepend"===t.location?r.prepend(t.view.el):"before"===t.location?r.before(t.view.el):"after"===t.location?r.after(t.view.el):r.append(t.view.el),t.view.onShow&&n.isFunction(t.view.onShow)&&t.view.onShow.apply(t.view,arguments),t.view.trigger("shown",this),t.view},getSubViewCount:function(){return this._subViews?this._subViews.length:0},removeSubViews:function(){n.each(this._subViews,function(n){this.stopListening(n.view),n.view.close()},this),this._subViews=[]},_removeSubView:function(t){var r=n.findWhere(this._subViews,{view:t}),e=this._subViews.indexOf(r);e>-1&&this._subViews.splice(e,1)},removeSubViewForModel:function(t){n.find(this._subViews,function(n){return n.view.model&&n.view.model===t?(n.view.close(),!0):void 0
-},this)},getAttr:function(n){return this.model?this.model.get(n):void 0},serializeForm:function(t){var r={},e=this.$(t).serializeArray();return n.each(e,function(n){r[n.name]?(r[n.name].push||(r[n.name]=[r[n.name]]),r[n.name].push(n.value||"")):r[n.name]=n.value||""}),r},close:function(){this.onClose&&n.isFunction(this.onClose)&&this.onClose.apply(this,arguments),this.removeSubViews(),this.remove(),this.trigger("closed",this),this.unbind()}})})},{backbone:22,underscore:21}],21:[function(n,t,r){arguments[4][19][0].apply(r,arguments)},{dup:19}],22:[function(n,t,r){!function(t,e){if("function"==typeof define&&define.amd)define(["underscore","jquery","exports"],function(n,r,i){t.Backbone=e(t,i,n,r)});else if("undefined"!=typeof r){var i=n("underscore");e(t,r,i)}else t.Backbone=e(t,{},t._,t.jQuery||t.Zepto||t.ender||t.$)}(this,function(n,t,r,e){{var i=n.Backbone,a=[],s=(a.push,a.slice);a.splice}t.VERSION="1.1.2",t.$=e,t.noConflict=function(){return n.Backbone=i,this},t.emulateHTTP=!1,t.emulateJSON=!1;var l=t.Events={on:function(n,t,r){if(!c(this,"on",n,[t,r])||!t)return this;this._events||(this._events={});var e=this._events[n]||(this._events[n]=[]);return e.push({callback:t,context:r,ctx:r||this}),this},once:function(n,t,e){if(!c(this,"once",n,[t,e])||!t)return this;var i=this,a=r.once(function(){i.off(n,a),t.apply(this,arguments)});return a._callback=t,this.on(n,a,e)},off:function(n,t,e){var i,a,s,l,o,u,d,f;if(!this._events||!c(this,"off",n,[t,e]))return this;if(!n&&!t&&!e)return this._events=void 0,this;for(l=n?[n]:r.keys(this._events),o=0,u=l.length;u>o;o++)if(n=l[o],s=this._events[n]){if(this._events[n]=i=[],t||e)for(d=0,f=s.length;f>d;d++)a=s[d],(t&&t!==a.callback&&t!==a.callback._callback||e&&e!==a.context)&&i.push(a);i.length||delete this._events[n]}return this},trigger:function(n){if(!this._events)return this;var t=s.call(arguments,1);if(!c(this,"trigger",n,t))return this;var r=this._events[n],e=this._events.all;return r&&u(r,t),e&&u(e,arguments),this},stopListening:function(n,t,e){var i=this._listeningTo;if(!i)return this;var a=!t&&!e;e||"object"!=typeof t||(e=this),n&&((i={})[n._listenId]=n);for(var s in i)n=i[s],n.off(t,e,this),(a||r.isEmpty(n._events))&&delete this._listeningTo[s];return this}},o=/\s+/,c=function(n,t,r,e){if(!r)return!0;if("object"==typeof r){for(var i in r)n[t].apply(n,[i,r[i]].concat(e));return!1}if(o.test(r)){for(var a=r.split(o),s=0,l=a.length;l>s;s++)n[t].apply(n,[a[s]].concat(e));return!1}return!0},u=function(n,t){var r,e=-1,i=n.length,a=t[0],s=t[1],l=t[2];switch(t.length){case 0:for(;++e<i;)(r=n[e]).callback.call(r.ctx);return;case 1:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a);return;case 2:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a,s);return;case 3:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a,s,l);return;default:for(;++e<i;)(r=n[e]).callback.apply(r.ctx,t);return}},d={listenTo:"on",listenToOnce:"once"};r.each(d,function(n,t){l[t]=function(t,e,i){var a=this._listeningTo||(this._listeningTo={}),s=t._listenId||(t._listenId=r.uniqueId("l"));return a[s]=t,i||"object"!=typeof e||(i=this),t[n](e,i,this),this}}),l.bind=l.on,l.unbind=l.off,r.extend(t,l);var f=t.Model=function(n,t){var e=n||{};t||(t={}),this.cid=r.uniqueId("c"),this.attributes={},t.collection&&(this.collection=t.collection),t.parse&&(e=this.parse(e,t)||{}),e=r.defaults({},e,r.result(this,"defaults")),this.set(e,t),this.changed={},this.initialize.apply(this,arguments)};r.extend(f.prototype,l,{changed:null,validationError:null,idAttribute:"id",initialize:function(){},toJSON:function(){return r.clone(this.attributes)},sync:function(){return t.sync.apply(this,arguments)},get:function(n){return this.attributes[n]},escape:function(n){return r.escape(this.get(n))},has:function(n){return null!=this.get(n)},set:function(n,t,e){var i,a,s,l,o,c,u,d;if(null==n)return this;if("object"==typeof n?(a=n,e=t):(a={})[n]=t,e||(e={}),!this._validate(a,e))return!1;s=e.unset,o=e.silent,l=[],c=this._changing,this._changing=!0,c||(this._previousAttributes=r.clone(this.attributes),this.changed={}),d=this.attributes,u=this._previousAttributes,this.idAttribute in a&&(this.id=a[this.idAttribute]);for(i in a)t=a[i],r.isEqual(d[i],t)||l.push(i),r.isEqual(u[i],t)?delete this.changed[i]:this.changed[i]=t,s?delete d[i]:d[i]=t;if(!o){l.length&&(this._pending=e);for(var f=0,h=l.length;h>f;f++)this.trigger("change:"+l[f],this,d[l[f]],e)}if(c)return this;if(!o)for(;this._pending;)e=this._pending,this._pending=!1,this.trigger("change",this,e);return this._pending=!1,this._changing=!1,this},unset:function(n,t){return this.set(n,void 0,r.extend({},t,{unset:!0}))},clear:function(n){var t={};for(var e in this.attributes)t[e]=void 0;return this.set(t,r.extend({},n,{unset:!0}))},hasChanged:function(n){return null==n?!r.isEmpty(this.changed):r.has(this.changed,n)},changedAttributes:function(n){if(!n)return this.hasChanged()?r.clone(this.changed):!1;var t,e=!1,i=this._changing?this._previousAttributes:this.attributes;for(var a in n)r.isEqual(i[a],t=n[a])||((e||(e={}))[a]=t);return e},previous:function(n){return null!=n&&this._previousAttributes?this._previousAttributes[n]:null},previousAttributes:function(){return r.clone(this._previousAttributes)},fetch:function(n){n=n?r.clone(n):{},void 0===n.parse&&(n.parse=!0);var t=this,e=n.success;return n.success=function(r){return t.set(t.parse(r,n),n)?(e&&e(t,r,n),void t.trigger("sync",t,r,n)):!1},J(this,n),this.sync("read",this,n)},save:function(n,t,e){var i,a,s,l=this.attributes;if(null==n||"object"==typeof n?(i=n,e=t):(i={})[n]=t,e=r.extend({validate:!0},e),i&&!e.wait){if(!this.set(i,e))return!1}else if(!this._validate(i,e))return!1;i&&e.wait&&(this.attributes=r.extend({},l,i)),void 0===e.parse&&(e.parse=!0);var o=this,c=e.success;return e.success=function(n){o.attributes=l;var t=o.parse(n,e);return e.wait&&(t=r.extend(i||{},t)),r.isObject(t)&&!o.set(t,e)?!1:(c&&c(o,n,e),void o.trigger("sync",o,n,e))},J(this,e),a=this.isNew()?"create":e.patch?"patch":"update","patch"===a&&(e.attrs=i),s=this.sync(a,this,e),i&&e.wait&&(this.attributes=l),s},destroy:function(n){n=n?r.clone(n):{};var t=this,e=n.success,i=function(){t.trigger("destroy",t,t.collection,n)};if(n.success=function(r){(n.wait||t.isNew())&&i(),e&&e(t,r,n),t.isNew()||t.trigger("sync",t,r,n)},this.isNew())return n.success(),!1;J(this,n);var a=this.sync("delete",this,n);return n.wait||i(),a},url:function(){var n=r.result(this,"urlRoot")||r.result(this.collection,"url")||V();return this.isNew()?n:n.replace(/([^\/])$/,"$1/")+encodeURIComponent(this.id)},parse:function(n){return n},clone:function(){return new this.constructor(this.attributes)},isNew:function(){return!this.has(this.idAttribute)},isValid:function(n){return this._validate({},r.extend(n||{},{validate:!0}))},_validate:function(n,t){if(!t.validate||!this.validate)return!0;n=r.extend({},this.attributes,n);var e=this.validationError=this.validate(n,t)||null;return e?(this.trigger("invalid",this,e,r.extend(t,{validationError:e})),!1):!0}});var h=["keys","values","pairs","invert","pick","omit"];r.each(h,function(n){f.prototype[n]=function(){var t=s.call(arguments);return t.unshift(this.attributes),r[n].apply(r,t)}});var p=t.Collection=function(n,t){t||(t={}),t.model&&(this.model=t.model),void 0!==t.comparator&&(this.comparator=t.comparator),this._reset(),this.initialize.apply(this,arguments),n&&this.reset(n,r.extend({silent:!0},t))},v={add:!0,remove:!0,merge:!0},m={add:!0,remove:!1};r.extend(p.prototype,l,{model:f,initialize:function(){},toJSON:function(n){return this.map(function(t){return t.toJSON(n)})},sync:function(){return t.sync.apply(this,arguments)},add:function(n,t){return this.set(n,r.extend({merge:!1},t,m))},remove:function(n,t){var e=!r.isArray(n);n=e?[n]:r.clone(n),t||(t={});var i,a,s,l;for(i=0,a=n.length;a>i;i++)l=n[i]=this.get(n[i]),l&&(delete this._byId[l.id],delete this._byId[l.cid],s=this.indexOf(l),this.models.splice(s,1),this.length--,t.silent||(t.index=s,l.trigger("remove",l,this,t)),this._removeReference(l,t));return e?n[0]:n},set:function(n,t){t=r.defaults({},t,v),t.parse&&(n=this.parse(n,t));var e=!r.isArray(n);n=e?n?[n]:[]:r.clone(n);var i,a,s,l,o,c,u,d=t.at,h=this.model,p=this.comparator&&null==d&&t.sort!==!1,m=r.isString(this.comparator)?this.comparator:null,g=[],b=[],y={},w=t.add,x=t.merge,_=t.remove,k=!p&&w&&_?[]:!1;for(i=0,a=n.length;a>i;i++){if(o=n[i]||{},s=o instanceof f?l=o:o[h.prototype.idAttribute||"id"],c=this.get(s))_&&(y[c.cid]=!0),x&&(o=o===l?l.attributes:o,t.parse&&(o=c.parse(o,t)),c.set(o,t),p&&!u&&c.hasChanged(m)&&(u=!0)),n[i]=c;else if(w){if(l=n[i]=this._prepareModel(o,t),!l)continue;g.push(l),this._addReference(l,t)}l=c||l,!k||!l.isNew()&&y[l.id]||k.push(l),y[l.id]=!0}if(_){for(i=0,a=this.length;a>i;++i)y[(l=this.models[i]).cid]||b.push(l);b.length&&this.remove(b,t)}if(g.length||k&&k.length)if(p&&(u=!0),this.length+=g.length,null!=d)for(i=0,a=g.length;a>i;i++)this.models.splice(d+i,0,g[i]);else{k&&(this.models.length=0);var j=k||g;for(i=0,a=j.length;a>i;i++)this.models.push(j[i])}if(u&&this.sort({silent:!0}),!t.silent){for(i=0,a=g.length;a>i;i++)(l=g[i]).trigger("add",l,this,t);(u||k&&k.length)&&this.trigger("sort",this,t)}return e?n[0]:n},reset:function(n,t){t||(t={});for(var e=0,i=this.models.length;i>e;e++)this._removeReference(this.models[e],t);return t.previousModels=this.models,this._reset(),n=this.add(n,r.extend({silent:!0},t)),t.silent||this.trigger("reset",this,t),n},push:function(n,t){return this.add(n,r.extend({at:this.length},t))},pop:function(n){var t=this.at(this.length-1);return this.remove(t,n),t},unshift:function(n,t){return this.add(n,r.extend({at:0},t))},shift:function(n){var t=this.at(0);return this.remove(t,n),t},slice:function(){return s.apply(this.models,arguments)},get:function(n){return null==n?void 0:this._byId[n]||this._byId[n.id]||this._byId[n.cid]},at:function(n){return this.models[n]},where:function(n,t){return r.isEmpty(n)?t?void 0:[]:this[t?"find":"filter"](function(t){for(var r in n)if(n[r]!==t.get(r))return!1;return!0})},findWhere:function(n){return this.where(n,!0)},sort:function(n){if(!this.comparator)throw new Error("Cannot sort a set without a comparator");return n||(n={}),r.isString(this.comparator)||1===this.comparator.length?this.models=this.sortBy(this.comparator,this):this.models.sort(r.bind(this.comparator,this)),n.silent||this.trigger("sort",this,n),this},pluck:function(n){return r.invoke(this.models,"get",n)},fetch:function(n){n=n?r.clone(n):{},void 0===n.parse&&(n.parse=!0);var t=n.success,e=this;return n.success=function(r){var i=n.reset?"reset":"set";e[i](r,n),t&&t(e,r,n),e.trigger("sync",e,r,n)},J(this,n),this.sync("read",this,n)},create:function(n,t){if(t=t?r.clone(t):{},!(n=this._prepareModel(n,t)))return!1;t.wait||this.add(n,t);var e=this,i=t.success;return t.success=function(n,r){t.wait&&e.add(n,t),i&&i(n,r,t)},n.save(null,t),n},parse:function(n){return n},clone:function(){return new this.constructor(this.models)},_reset:function(){this.length=0,this.models=[],this._byId={}},_prepareModel:function(n,t){if(n instanceof f)return n;t=t?r.clone(t):{},t.collection=this;var e=new this.model(n,t);return e.validationError?(this.trigger("invalid",this,e.validationError,t),!1):e},_addReference:function(n){this._byId[n.cid]=n,null!=n.id&&(this._byId[n.id]=n),n.collection||(n.collection=this),n.on("all",this._onModelEvent,this)},_removeReference:function(n){this===n.collection&&delete n.collection,n.off("all",this._onModelEvent,this)},_onModelEvent:function(n,t,r,e){("add"!==n&&"remove"!==n||r===this)&&("destroy"===n&&this.remove(t,e),t&&n==="change:"+t.idAttribute&&(delete this._byId[t.previous(t.idAttribute)],null!=t.id&&(this._byId[t.id]=t)),this.trigger.apply(this,arguments))}});var g=["forEach","each","map","collect","reduce","foldl","inject","reduceRight","foldr","find","detect","filter","select","reject","every","all","some","any","include","contains","invoke","max","min","toArray","size","first","head","take","initial","rest","tail","drop","last","without","difference","indexOf","shuffle","lastIndexOf","isEmpty","chain","sample"];r.each(g,function(n){p.prototype[n]=function(){var t=s.call(arguments);return t.unshift(this.models),r[n].apply(r,t)}});var b=["groupBy","countBy","sortBy","indexBy"];r.each(b,function(n){p.prototype[n]=function(t,e){var i=r.isFunction(t)?t:function(n){return n.get(t)};return r[n](this.models,i,e)}});var y=t.View=function(n){this.cid=r.uniqueId("view"),n||(n={}),r.extend(this,r.pick(n,x)),this._ensureElement(),this.initialize.apply(this,arguments),this.delegateEvents()},w=/^(\S+)\s*(.*)$/,x=["model","collection","el","id","attributes","className","tagName","events"];r.extend(y.prototype,l,{tagName:"div",$:function(n){return this.$el.find(n)},initialize:function(){},render:function(){return this},remove:function(){return this.$el.remove(),this.stopListening(),this},setElement:function(n,r){return this.$el&&this.undelegateEvents(),this.$el=n instanceof t.$?n:t.$(n),this.el=this.$el[0],r!==!1&&this.delegateEvents(),this},delegateEvents:function(n){if(!n&&!(n=r.result(this,"events")))return this;this.undelegateEvents();for(var t in n){var e=n[t];if(r.isFunction(e)||(e=this[n[t]]),e){var i=t.match(w),a=i[1],s=i[2];e=r.bind(e,this),a+=".delegateEvents"+this.cid,""===s?this.$el.on(a,e):this.$el.on(a,s,e)}}return this},undelegateEvents:function(){return this.$el.off(".delegateEvents"+this.cid),this},_ensureElement:function(){if(this.el)this.setElement(r.result(this,"el"),!1);else{var n=r.extend({},r.result(this,"attributes"));this.id&&(n.id=r.result(this,"id")),this.className&&(n["class"]=r.result(this,"className"));var e=t.$("<"+r.result(this,"tagName")+">").attr(n);this.setElement(e,!1)}}}),t.sync=function(n,e,i){var a=k[n];r.defaults(i||(i={}),{emulateHTTP:t.emulateHTTP,emulateJSON:t.emulateJSON});var s={type:a,dataType:"json"};if(i.url||(s.url=r.result(e,"url")||V()),null!=i.data||!e||"create"!==n&&"update"!==n&&"patch"!==n||(s.contentType="application/json",s.data=JSON.stringify(i.attrs||e.toJSON(i))),i.emulateJSON&&(s.contentType="application/x-www-form-urlencoded",s.data=s.data?{model:s.data}:{}),i.emulateHTTP&&("PUT"===a||"DELETE"===a||"PATCH"===a)){s.type="POST",i.emulateJSON&&(s.data._method=a);var l=i.beforeSend;i.beforeSend=function(n){return n.setRequestHeader("X-HTTP-Method-Override",a),l?l.apply(this,arguments):void 0}}"GET"===s.type||i.emulateJSON||(s.processData=!1),"PATCH"===s.type&&_&&(s.xhr=function(){return new ActiveXObject("Microsoft.XMLHTTP")});var o=i.xhr=t.ajax(r.extend(s,i));return e.trigger("request",e,o,i),o};var _=!("undefined"==typeof window||!window.ActiveXObject||window.XMLHttpRequest&&(new XMLHttpRequest).dispatchEvent),k={create:"POST",update:"PUT",patch:"PATCH","delete":"DELETE",read:"GET"};t.ajax=function(){return t.$.ajax.apply(t.$,arguments)};var j=t.Router=function(n){n||(n={}),n.routes&&(this.routes=n.routes),this._bindRoutes(),this.initialize.apply(this,arguments)},A=/\((.*?)\)/g,S=/(\(\?)?:\w+/g,E=/\*\w+/g,T=/[\-{}\[\]+?.,\\\^$|#\s]/g;r.extend(j.prototype,l,{initialize:function(){},route:function(n,e,i){r.isRegExp(n)||(n=this._routeToRegExp(n)),r.isFunction(e)&&(i=e,e=""),i||(i=this[e]);var a=this;return t.history.route(n,function(r){var s=a._extractParameters(n,r);a.execute(i,s),a.trigger.apply(a,["route:"+e].concat(s)),a.trigger("route",e,s),t.history.trigger("route",a,e,s)}),this},execute:function(n,t){n&&n.apply(this,t)},navigate:function(n,r){return t.history.navigate(n,r),this},_bindRoutes:function(){if(this.routes){this.routes=r.result(this,"routes");for(var n,t=r.keys(this.routes);null!=(n=t.pop());)this.route(n,this.routes[n])}},_routeToRegExp:function(n){return n=n.replace(T,"\\$&").replace(A,"(?:$1)?").replace(S,function(n,t){return t?n:"([^/?]+)"}).replace(E,"([^?]*?)"),new RegExp("^"+n+"(?:\\?([\\s\\S]*))?$")},_extractParameters:function(n,t){var e=n.exec(t).slice(1);return r.map(e,function(n,t){return t===e.length-1?n||null:n?decodeURIComponent(n):null})}});var O=t.History=function(){this.handlers=[],r.bindAll(this,"checkUrl"),"undefined"!=typeof window&&(this.location=window.location,this.history=window.history)},C=/^[#\/]|\s+$/g,$=/^\/+|\/+$/g,F=/msie [\w.]+/,N=/\/$/,M=/#.*$/;O.started=!1,r.extend(O.prototype,l,{interval:50,atRoot:function(){return this.location.pathname.replace(/[^\/]$/,"$&/")===this.root},getHash:function(n){var t=(n||this).location.href.match(/#(.*)$/);return t?t[1]:""},getFragment:function(n,t){if(null==n)if(this._hasPushState||!this._wantsHashChange||t){n=decodeURI(this.location.pathname+this.location.search);var r=this.root.replace(N,"");n.indexOf(r)||(n=n.slice(r.length))}else n=this.getHash();return n.replace(C,"")},start:function(n){if(O.started)throw new Error("Backbone.history has already been started");O.started=!0,this.options=r.extend({root:"/"},this.options,n),this.root=this.options.root,this._wantsHashChange=this.options.hashChange!==!1,this._wantsPushState=!!this.options.pushState,this._hasPushState=!!(this.options.pushState&&this.history&&this.history.pushState);var e=this.getFragment(),i=document.documentMode,a=F.exec(navigator.userAgent.toLowerCase())&&(!i||7>=i);if(this.root=("/"+this.root+"/").replace($,"/"),a&&this._wantsHashChange){var s=t.$('<iframe src="javascript:0" tabindex="-1">');this.iframe=s.hide().appendTo("body")[0].contentWindow,this.navigate(e)}this._hasPushState?t.$(window).on("popstate",this.checkUrl):this._wantsHashChange&&"onhashchange"in window&&!a?t.$(window).on("hashchange",this.checkUrl):this._wantsHashChange&&(this._checkUrlInterval=setInterval(this.checkUrl,this.interval)),this.fragment=e;var l=this.location;if(this._wantsHashChange&&this._wantsPushState){if(!this._hasPushState&&!this.atRoot())return this.fragment=this.getFragment(null,!0),this.location.replace(this.root+"#"+this.fragment),!0;this._hasPushState&&this.atRoot()&&l.hash&&(this.fragment=this.getHash().replace(C,""),this.history.replaceState({},document.title,this.root+this.fragment))}return this.options.silent?void 0:this.loadUrl()},stop:function(){t.$(window).off("popstate",this.checkUrl).off("hashchange",this.checkUrl),this._checkUrlInterval&&clearInterval(this._checkUrlInterval),O.started=!1},route:function(n,t){this.handlers.unshift({route:n,callback:t})},checkUrl:function(){var n=this.getFragment();return n===this.fragment&&this.iframe&&(n=this.getFragment(this.getHash(this.iframe))),n===this.fragment?!1:(this.iframe&&this.navigate(n),void this.loadUrl())},loadUrl:function(n){return n=this.fragment=this.getFragment(n),r.any(this.handlers,function(t){return t.route.test(n)?(t.callback(n),!0):void 0})},navigate:function(n,t){if(!O.started)return!1;t&&t!==!0||(t={trigger:!!t});var r=this.root+(n=this.getFragment(n||""));if(n=n.replace(M,""),this.fragment!==n){if(this.fragment=n,""===n&&"/"!==r&&(r=r.slice(0,-1)),this._hasPushState)this.history[t.replace?"replaceState":"pushState"]({},document.title,r);else{if(!this._wantsHashChange)return this.location.assign(r);this._updateHash(this.location,n,t.replace),this.iframe&&n!==this.getFragment(this.getHash(this.iframe))&&(t.replace||this.iframe.document.open().close(),this._updateHash(this.iframe.location,n,t.replace))}return t.trigger?this.loadUrl(n):void 0}},_updateHash:function(n,t,r){if(r){var e=n.href.replace(/(javascript:|#).*$/,"");n.replace(e+"#"+t)}else n.hash="#"+t}}),t.history=new O;var I=function(n,t){var e,i=this;e=n&&r.has(n,"constructor")?n.constructor:function(){return i.apply(this,arguments)},r.extend(e,i,t);var a=function(){this.constructor=e};return a.prototype=i.prototype,e.prototype=new a,n&&r.extend(e.prototype,n),e.__super__=i.prototype,e};f.extend=p.extend=j.extend=y.extend=O.extend=I;var V=function(){throw new Error('A "url" property or function must be specified')},J=function(n,t){var r=t.error;t.error=function(e){r&&r(n,e,t),n.trigger("error",n,e,t)}};return t})},{underscore:23}],23:[function(n,t,r){(function(){function n(n){function t(t,r,e,i,a,s){for(;a>=0&&s>a;a+=n){var l=i?i[a]:a;e=r(e,t[l],l,t)}return e}return function(r,e,i,a){e=w(e,a,4);var s=!A(r)&&y.keys(r),l=(s||r).length,o=n>0?0:l-1;return arguments.length<3&&(i=r[s?s[o]:o],o+=n),t(r,e,i,s,o,l)}}function e(n){return function(t,r,e){r=x(r,e);for(var i=null!=t&&t.length,a=n>0?0:i-1;a>=0&&i>a;a+=n)if(r(t[a],a,t))return a;return-1}}function i(n,t){var r=C.length,e=n.constructor,i=y.isFunction(e)&&e.prototype||o,a="constructor";for(y.has(n,a)&&!y.contains(t,a)&&t.push(a);r--;)a=C[r],a in n&&n[a]!==i[a]&&!y.contains(t,a)&&t.push(a)}var a=this,s=a._,l=Array.prototype,o=Object.prototype,c=Function.prototype,u=l.push,d=l.slice,f=o.toString,h=o.hasOwnProperty,p=Array.isArray,v=Object.keys,m=c.bind,g=Object.create,b=function(){},y=function(n){return n instanceof y?n:this instanceof y?void(this._wrapped=n):new y(n)};"undefined"!=typeof r?("undefined"!=typeof t&&t.exports&&(r=t.exports=y),r._=y):a._=y,y.VERSION="1.8.2";var w=function(n,t,r){if(void 0===t)return n;switch(null==r?3:r){case 1:return function(r){return n.call(t,r)};case 2:return function(r,e){return n.call(t,r,e)};case 3:return function(r,e,i){return n.call(t,r,e,i)};case 4:return function(r,e,i,a){return n.call(t,r,e,i,a)}}return function(){return n.apply(t,arguments)}},x=function(n,t,r){return null==n?y.identity:y.isFunction(n)?w(n,t,r):y.isObject(n)?y.matcher(n):y.property(n)};y.iteratee=function(n,t){return x(n,t,1/0)};var _=function(n,t){return function(r){var e=arguments.length;if(2>e||null==r)return r;for(var i=1;e>i;i++)for(var a=arguments[i],s=n(a),l=s.length,o=0;l>o;o++){var c=s[o];t&&void 0!==r[c]||(r[c]=a[c])}return r}},k=function(n){if(!y.isObject(n))return{};if(g)return g(n);b.prototype=n;var t=new b;return b.prototype=null,t},j=Math.pow(2,53)-1,A=function(n){var t=n&&n.length;return"number"==typeof t&&t>=0&&j>=t};y.each=y.forEach=function(n,t,r){t=w(t,r);var e,i;if(A(n))for(e=0,i=n.length;i>e;e++)t(n[e],e,n);else{var a=y.keys(n);for(e=0,i=a.length;i>e;e++)t(n[a[e]],a[e],n)}return n},y.map=y.collect=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=Array(i),s=0;i>s;s++){var l=e?e[s]:s;a[s]=t(n[l],l,n)}return a},y.reduce=y.foldl=y.inject=n(1),y.reduceRight=y.foldr=n(-1),y.find=y.detect=function(n,t,r){var e;return e=A(n)?y.findIndex(n,t,r):y.findKey(n,t,r),void 0!==e&&-1!==e?n[e]:void 0},y.filter=y.select=function(n,t,r){var e=[];return t=x(t,r),y.each(n,function(n,r,i){t(n,r,i)&&e.push(n)}),e},y.reject=function(n,t,r){return y.filter(n,y.negate(x(t)),r)},y.every=y.all=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=0;i>a;a++){var s=e?e[a]:a;if(!t(n[s],s,n))return!1}return!0},y.some=y.any=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=0;i>a;a++){var s=e?e[a]:a;if(t(n[s],s,n))return!0}return!1},y.contains=y.includes=y.include=function(n,t,r){return A(n)||(n=y.values(n)),y.indexOf(n,t,"number"==typeof r&&r)>=0},y.invoke=function(n,t){var r=d.call(arguments,2),e=y.isFunction(t);return y.map(n,function(n){var i=e?t:n[t];return null==i?i:i.apply(n,r)})},y.pluck=function(n,t){return y.map(n,y.property(t))},y.where=function(n,t){return y.filter(n,y.matcher(t))},y.findWhere=function(n,t){return y.find(n,y.matcher(t))},y.max=function(n,t,r){var e,i,a=-1/0,s=-1/0;if(null==t&&null!=n){n=A(n)?n:y.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],e>a&&(a=e)}else t=x(t,r),y.each(n,function(n,r,e){i=t(n,r,e),(i>s||i===-1/0&&a===-1/0)&&(a=n,s=i)});return a},y.min=function(n,t,r){var e,i,a=1/0,s=1/0;if(null==t&&null!=n){n=A(n)?n:y.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],a>e&&(a=e)}else t=x(t,r),y.each(n,function(n,r,e){i=t(n,r,e),(s>i||1/0===i&&1/0===a)&&(a=n,s=i)});return a},y.shuffle=function(n){for(var t,r=A(n)?n:y.values(n),e=r.length,i=Array(e),a=0;e>a;a++)t=y.random(0,a),t!==a&&(i[a]=i[t]),i[t]=r[a];return i},y.sample=function(n,t,r){return null==t||r?(A(n)||(n=y.values(n)),n[y.random(n.length-1)]):y.shuffle(n).slice(0,Math.max(0,t))},y.sortBy=function(n,t,r){return t=x(t,r),y.pluck(y.map(n,function(n,r,e){return{value:n,index:r,criteria:t(n,r,e)}}).sort(function(n,t){var r=n.criteria,e=t.criteria;if(r!==e){if(r>e||void 0===r)return 1;if(e>r||void 0===e)return-1}return n.index-t.index}),"value")};var S=function(n){return function(t,r,e){var i={};return r=x(r,e),y.each(t,function(e,a){var s=r(e,a,t);n(i,e,s)}),i}};y.groupBy=S(function(n,t,r){y.has(n,r)?n[r].push(t):n[r]=[t]}),y.indexBy=S(function(n,t,r){n[r]=t}),y.countBy=S(function(n,t,r){y.has(n,r)?n[r]++:n[r]=1}),y.toArray=function(n){return n?y.isArray(n)?d.call(n):A(n)?y.map(n,y.identity):y.values(n):[]},y.size=function(n){return null==n?0:A(n)?n.length:y.keys(n).length},y.partition=function(n,t,r){t=x(t,r);var e=[],i=[];return y.each(n,function(n,r,a){(t(n,r,a)?e:i).push(n)}),[e,i]},y.first=y.head=y.take=function(n,t,r){return null==n?void 0:null==t||r?n[0]:y.initial(n,n.length-t)},y.initial=function(n,t,r){return d.call(n,0,Math.max(0,n.length-(null==t||r?1:t)))},y.last=function(n,t,r){return null==n?void 0:null==t||r?n[n.length-1]:y.rest(n,Math.max(0,n.length-t))},y.rest=y.tail=y.drop=function(n,t,r){return d.call(n,null==t||r?1:t)},y.compact=function(n){return y.filter(n,y.identity)};var E=function(n,t,r,e){for(var i=[],a=0,s=e||0,l=n&&n.length;l>s;s++){var o=n[s];if(A(o)&&(y.isArray(o)||y.isArguments(o))){t||(o=E(o,t,r));var c=0,u=o.length;for(i.length+=u;u>c;)i[a++]=o[c++]}else r||(i[a++]=o)}return i};y.flatten=function(n,t){return E(n,t,!1)},y.without=function(n){return y.difference(n,d.call(arguments,1))},y.uniq=y.unique=function(n,t,r,e){if(null==n)return[];y.isBoolean(t)||(e=r,r=t,t=!1),null!=r&&(r=x(r,e));for(var i=[],a=[],s=0,l=n.length;l>s;s++){var o=n[s],c=r?r(o,s,n):o;t?(s&&a===c||i.push(o),a=c):r?y.contains(a,c)||(a.push(c),i.push(o)):y.contains(i,o)||i.push(o)}return i},y.union=function(){return y.uniq(E(arguments,!0,!0))},y.intersection=function(n){if(null==n)return[];for(var t=[],r=arguments.length,e=0,i=n.length;i>e;e++){var a=n[e];if(!y.contains(t,a)){for(var s=1;r>s&&y.contains(arguments[s],a);s++);s===r&&t.push(a)}}return t},y.difference=function(n){var t=E(arguments,!0,!0,1);return y.filter(n,function(n){return!y.contains(t,n)})},y.zip=function(){return y.unzip(arguments)},y.unzip=function(n){for(var t=n&&y.max(n,"length").length||0,r=Array(t),e=0;t>e;e++)r[e]=y.pluck(n,e);return r},y.object=function(n,t){for(var r={},e=0,i=n&&n.length;i>e;e++)t?r[n[e]]=t[e]:r[n[e][0]]=n[e][1];return r},y.indexOf=function(n,t,r){var e=0,i=n&&n.length;if("number"==typeof r)e=0>r?Math.max(0,i+r):r;else if(r&&i)return e=y.sortedIndex(n,t),n[e]===t?e:-1;if(t!==t)return y.findIndex(d.call(n,e),y.isNaN);for(;i>e;e++)if(n[e]===t)return e;return-1},y.lastIndexOf=function(n,t,r){var e=n?n.length:0;if("number"==typeof r&&(e=0>r?e+r+1:Math.min(e,r+1)),t!==t)return y.findLastIndex(d.call(n,0,e),y.isNaN);for(;--e>=0;)if(n[e]===t)return e;return-1},y.findIndex=e(1),y.findLastIndex=e(-1),y.sortedIndex=function(n,t,r,e){r=x(r,e,1);for(var i=r(t),a=0,s=n.length;s>a;){var l=Math.floor((a+s)/2);r(n[l])<i?a=l+1:s=l}return a},y.range=function(n,t,r){arguments.length<=1&&(t=n||0,n=0),r=r||1;for(var e=Math.max(Math.ceil((t-n)/r),0),i=Array(e),a=0;e>a;a++,n+=r)i[a]=n;return i};var T=function(n,t,r,e,i){if(!(e instanceof t))return n.apply(r,i);var a=k(n.prototype),s=n.apply(a,i);return y.isObject(s)?s:a};y.bind=function(n,t){if(m&&n.bind===m)return m.apply(n,d.call(arguments,1));if(!y.isFunction(n))throw new TypeError("Bind must be called on a function");var r=d.call(arguments,2),e=function(){return T(n,e,t,this,r.concat(d.call(arguments)))};return e},y.partial=function(n){var t=d.call(arguments,1),r=function(){for(var e=0,i=t.length,a=Array(i),s=0;i>s;s++)a[s]=t[s]===y?arguments[e++]:t[s];for(;e<arguments.length;)a.push(arguments[e++]);return T(n,r,this,this,a)};return r},y.bindAll=function(n){var t,r,e=arguments.length;if(1>=e)throw new Error("bindAll must be passed function names");for(t=1;e>t;t++)r=arguments[t],n[r]=y.bind(n[r],n);return n},y.memoize=function(n,t){var r=function(e){var i=r.cache,a=""+(t?t.apply(this,arguments):e);return y.has(i,a)||(i[a]=n.apply(this,arguments)),i[a]};return r.cache={},r},y.delay=function(n,t){var r=d.call(arguments,2);return setTimeout(function(){return n.apply(null,r)},t)},y.defer=y.partial(y.delay,y,1),y.throttle=function(n,t,r){var e,i,a,s=null,l=0;r||(r={});var o=function(){l=r.leading===!1?0:y.now(),s=null,a=n.apply(e,i),s||(e=i=null)};return function(){var c=y.now();l||r.leading!==!1||(l=c);var u=t-(c-l);return e=this,i=arguments,0>=u||u>t?(s&&(clearTimeout(s),s=null),l=c,a=n.apply(e,i),s||(e=i=null)):s||r.trailing===!1||(s=setTimeout(o,u)),a}},y.debounce=function(n,t,r){var e,i,a,s,l,o=function(){var c=y.now()-s;t>c&&c>=0?e=setTimeout(o,t-c):(e=null,r||(l=n.apply(a,i),e||(a=i=null)))};return function(){a=this,i=arguments,s=y.now();var c=r&&!e;return e||(e=setTimeout(o,t)),c&&(l=n.apply(a,i),a=i=null),l}},y.wrap=function(n,t){return y.partial(t,n)},y.negate=function(n){return function(){return!n.apply(this,arguments)}},y.compose=function(){var n=arguments,t=n.length-1;return function(){for(var r=t,e=n[t].apply(this,arguments);r--;)e=n[r].call(this,e);return e}},y.after=function(n,t){return function(){return--n<1?t.apply(this,arguments):void 0}},y.before=function(n,t){var r;return function(){return--n>0&&(r=t.apply(this,arguments)),1>=n&&(t=null),r}},y.once=y.partial(y.before,2);var O=!{toString:null}.propertyIsEnumerable("toString"),C=["valueOf","isPrototypeOf","toString","propertyIsEnumerable","hasOwnProperty","toLocaleString"];y.keys=function(n){if(!y.isObject(n))return[];if(v)return v(n);var t=[];for(var r in n)y.has(n,r)&&t.push(r);return O&&i(n,t),t},y.allKeys=function(n){if(!y.isObject(n))return[];var t=[];for(var r in n)t.push(r);return O&&i(n,t),t},y.values=function(n){for(var t=y.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=n[t[i]];return e},y.mapObject=function(n,t,r){t=x(t,r);for(var e,i=y.keys(n),a=i.length,s={},l=0;a>l;l++)e=i[l],s[e]=t(n[e],e,n);return s},y.pairs=function(n){for(var t=y.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=[t[i],n[t[i]]];return e},y.invert=function(n){for(var t={},r=y.keys(n),e=0,i=r.length;i>e;e++)t[n[r[e]]]=r[e];return t},y.functions=y.methods=function(n){var t=[];for(var r in n)y.isFunction(n[r])&&t.push(r);return t.sort()},y.extend=_(y.allKeys),y.extendOwn=y.assign=_(y.keys),y.findKey=function(n,t,r){t=x(t,r);for(var e,i=y.keys(n),a=0,s=i.length;s>a;a++)if(e=i[a],t(n[e],e,n))return e},y.pick=function(n,t,r){var e,i,a={},s=n;if(null==s)return a;y.isFunction(t)?(i=y.allKeys(s),e=w(t,r)):(i=E(arguments,!1,!1,1),e=function(n,t,r){return t in r},s=Object(s));for(var l=0,o=i.length;o>l;l++){var c=i[l],u=s[c];e(u,c,s)&&(a[c]=u)}return a},y.omit=function(n,t,r){if(y.isFunction(t))t=y.negate(t);else{var e=y.map(E(arguments,!1,!1,1),String);t=function(n,t){return!y.contains(e,t)}}return y.pick(n,t,r)},y.defaults=_(y.allKeys,!0),y.clone=function(n){return y.isObject(n)?y.isArray(n)?n.slice():y.extend({},n):n},y.tap=function(n,t){return t(n),n},y.isMatch=function(n,t){var r=y.keys(t),e=r.length;if(null==n)return!e;for(var i=Object(n),a=0;e>a;a++){var s=r[a];if(t[s]!==i[s]||!(s in i))return!1}return!0};var $=function(n,t,r,e){if(n===t)return 0!==n||1/n===1/t;if(null==n||null==t)return n===t;n instanceof y&&(n=n._wrapped),t instanceof y&&(t=t._wrapped);var i=f.call(n);if(i!==f.call(t))return!1;switch(i){case"[object RegExp]":case"[object String]":return""+n==""+t;case"[object Number]":return+n!==+n?+t!==+t:0===+n?1/+n===1/t:+n===+t;case"[object Date]":case"[object Boolean]":return+n===+t}var a="[object Array]"===i;if(!a){if("object"!=typeof n||"object"!=typeof t)return!1;var s=n.constructor,l=t.constructor;if(s!==l&&!(y.isFunction(s)&&s instanceof s&&y.isFunction(l)&&l instanceof l)&&"constructor"in n&&"constructor"in t)return!1}r=r||[],e=e||[];for(var o=r.length;o--;)if(r[o]===n)return e[o]===t;if(r.push(n),e.push(t),a){if(o=n.length,o!==t.length)return!1;for(;o--;)if(!$(n[o],t[o],r,e))return!1}else{var c,u=y.keys(n);if(o=u.length,y.keys(t).length!==o)return!1;for(;o--;)if(c=u[o],!y.has(t,c)||!$(n[c],t[c],r,e))return!1}return r.pop(),e.pop(),!0
-};y.isEqual=function(n,t){return $(n,t)},y.isEmpty=function(n){return null==n?!0:A(n)&&(y.isArray(n)||y.isString(n)||y.isArguments(n))?0===n.length:0===y.keys(n).length},y.isElement=function(n){return!(!n||1!==n.nodeType)},y.isArray=p||function(n){return"[object Array]"===f.call(n)},y.isObject=function(n){var t=typeof n;return"function"===t||"object"===t&&!!n},y.each(["Arguments","Function","String","Number","Date","RegExp","Error"],function(n){y["is"+n]=function(t){return f.call(t)==="[object "+n+"]"}}),y.isArguments(arguments)||(y.isArguments=function(n){return y.has(n,"callee")}),"function"!=typeof/./&&"object"!=typeof Int8Array&&(y.isFunction=function(n){return"function"==typeof n||!1}),y.isFinite=function(n){return isFinite(n)&&!isNaN(parseFloat(n))},y.isNaN=function(n){return y.isNumber(n)&&n!==+n},y.isBoolean=function(n){return n===!0||n===!1||"[object Boolean]"===f.call(n)},y.isNull=function(n){return null===n},y.isUndefined=function(n){return void 0===n},y.has=function(n,t){return null!=n&&h.call(n,t)},y.noConflict=function(){return a._=s,this},y.identity=function(n){return n},y.constant=function(n){return function(){return n}},y.noop=function(){},y.property=function(n){return function(t){return null==t?void 0:t[n]}},y.propertyOf=function(n){return null==n?function(){}:function(t){return n[t]}},y.matcher=y.matches=function(n){return n=y.extendOwn({},n),function(t){return y.isMatch(t,n)}},y.times=function(n,t,r){var e=Array(Math.max(0,n));t=w(t,r,1);for(var i=0;n>i;i++)e[i]=t(i);return e},y.random=function(n,t){return null==t&&(t=n,n=0),n+Math.floor(Math.random()*(t-n+1))},y.now=Date.now||function(){return(new Date).getTime()};var F={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#x27;","`":"&#x60;"},N=y.invert(F),M=function(n){var t=function(t){return n[t]},r="(?:"+y.keys(n).join("|")+")",e=RegExp(r),i=RegExp(r,"g");return function(n){return n=null==n?"":""+n,e.test(n)?n.replace(i,t):n}};y.escape=M(F),y.unescape=M(N),y.result=function(n,t,r){var e=null==n?void 0:n[t];return void 0===e&&(e=r),y.isFunction(e)?e.call(n):e};var I=0;y.uniqueId=function(n){var t=++I+"";return n?n+t:t},y.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g,escape:/<%-([\s\S]+?)%>/g};var V=/(.)^/,J={"'":"'","\\":"\\","\r":"r","\n":"n","\u2028":"u2028","\u2029":"u2029"},R=/\\|'|\r|\n|\u2028|\u2029/g,P=function(n){return"\\"+J[n]};y.template=function(n,t,r){!t&&r&&(t=r),t=y.defaults({},t,y.templateSettings);var e=RegExp([(t.escape||V).source,(t.interpolate||V).source,(t.evaluate||V).source].join("|")+"|$","g"),i=0,a="__p+='";n.replace(e,function(t,r,e,s,l){return a+=n.slice(i,l).replace(R,P),i=l+t.length,r?a+="'+\n((__t=("+r+"))==null?'':_.escape(__t))+\n'":e?a+="'+\n((__t=("+e+"))==null?'':__t)+\n'":s&&(a+="';\n"+s+"\n__p+='"),t}),a+="';\n",t.variable||(a="with(obj||{}){\n"+a+"}\n"),a="var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n"+a+"return __p;\n";try{var s=new Function(t.variable||"obj","_",a)}catch(l){throw l.source=a,l}var o=function(n){return s.call(this,n,y)},c=t.variable||"obj";return o.source="function("+c+"){\n"+a+"}",o},y.chain=function(n){var t=y(n);return t._chain=!0,t};var q=function(n,t){return n._chain?y(t).chain():t};y.mixin=function(n){y.each(y.functions(n),function(t){var r=y[t]=n[t];y.prototype[t]=function(){var n=[this._wrapped];return u.apply(n,arguments),q(this,r.apply(y,n))}})},y.mixin(y),y.each(["pop","push","reverse","shift","sort","splice","unshift"],function(n){var t=l[n];y.prototype[n]=function(){var r=this._wrapped;return t.apply(r,arguments),"shift"!==n&&"splice"!==n||0!==r.length||delete r[0],q(this,r)}}),y.each(["concat","join","slice"],function(n){var t=l[n];y.prototype[n]=function(){return q(this,t.apply(this._wrapped,arguments))}}),y.prototype.value=function(){return this._wrapped},y.prototype.valueOf=y.prototype.toJSON=y.prototype.value,y.prototype.toString=function(){return""+this._wrapped},"function"==typeof define&&define.amd&&define("underscore",[],function(){return y})}).call(this)},{}],24:[function(n,t,r){arguments[4][23][0].apply(r,arguments)},{dup:23}]},{},[1]);
+!function n(t,r,e){function i(s,l){if(!r[s]){if(!t[s]){var o="function"==typeof require&&require;if(!l&&o)return o(s,!0);if(a)return a(s,!0);var c=new Error("Cannot find module '"+s+"'");throw c.code="MODULE_NOT_FOUND",c}var u=r[s]={exports:{}};t[s][0].call(u.exports,function(n){var r=t[s][1][n];return i(r?r:n)},u,u.exports,n,t,r,e)}return r[s].exports}for(var a="function"==typeof require&&require,s=0;s<e.length;s++)i(e[s]);return i}({1:[function(n){var t=n("backbone"),r=(n("backbone.composer"),n("backbone.broker")),e=n("./router"),i=n("./views/main");t.$=window.$,t.View.prototype.attachToTemplate=!0,$(document).ready(function(){new i({el:"body"}),new e;r.start(),t.history.start()})},{"./router":3,"./views/main":21,backbone:28,"backbone.broker":24,"backbone.composer":26}],2:[function(n,t){var r=n("backbone");t.exports=r.Model.extend({defaults:{name:"David Williams"}})},{backbone:28}],3:[function(n,t){var r=n("backbone"),e=n("backbone.broker"),i=n("./views/dashboard"),a=n("./views/file-upload"),s=n("./views/file-manager"),l=e.channel("container"),o=e.channel("nav");t.exports=r.Router.extend({routes:{"":function(){l.publish("show",new i),o.publish("select","dashboard")},"file-manager":function(){l.publish("show",new s),o.publish("select","file-manager")},"file-upload":function(){l.publish("show",new a),o.publish("select","file-manager")}}})},{"./views/dashboard":14,"./views/file-manager":18,"./views/file-upload":19,backbone:28,"backbone.broker":24}],4:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>Dashboard</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content">\r\n  <div class="row">\r\n    <div class="col-md-2">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-success pull-right">Monthly</span>\r\n          <h5>Views</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <h1 class="no-margins">386,200</h1>\r\n          <div class="stat-percent font-bold text-success">98% <i class="fa fa-bolt"></i></div>\r\n          <small>Total views</small>\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-md-2">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-info pull-right">Annual</span>\r\n          <h5>Orders</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <h1 class="no-margins">80,800</h1>\r\n          <div class="stat-percent font-bold text-info">20% <i class="fa fa-level-up"></i></div>\r\n          <small>New orders</small>\r\n        </div>\r\n      </div>\r\n    </div>\r\n\r\n    <div class="col-md-4">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-primary pull-right">Today</span>\r\n          <h5>Vistits</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n\r\n          <div class="row">\r\n            <div class="col-md-6">\r\n              <h1 class="no-margins">$ 406,420</h1>\r\n              <div class="font-bold text-navy">44% <i class="fa fa-level-up"></i> <small>Rapid pace</small></div>\r\n            </div>\r\n            <div class="col-md-6">\r\n              <h1 class="no-margins">206,120</h1>\r\n              <div class="font-bold text-navy">22% <i class="fa fa-level-up"></i> <small>Slow pace</small></div>\r\n            </div>\r\n          </div>\r\n\r\n\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-md-4" id="flot-chart">\r\n    </div>\r\n  </div>\r\n  <div class="row">\r\n    <div class="col-lg-8" id="chart-js">\r\n    </div>\r\n    <div class="col-lg-4">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <span class="label label-warning pull-right">Data has changed</span>\r\n          <h5>User activity</h5>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>236 321.80</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>46.11%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>432.021</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>643 321.10</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>92.43%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>564.554</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <div class="row">\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Pages / Visit</small>\r\n              <h4>436 547.20</h4>\r\n            </div>\r\n\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">% New Visits</small>\r\n              <h4>150.23%</h4>\r\n            </div>\r\n            <div class="col-xs-4">\r\n              <small class="stats-label">Last week</small>\r\n              <h4>124.990</h4>\r\n            </div>\r\n          </div>\r\n        </div>\r\n      </div>\r\n    </div>\r\n\r\n  </div>\r\n\r\n  <div class="row">\r\n    <div class="col-lg-12" id="table">\r\n    </div>\r\n  </div>\r\n';return __p}},{underscore:30}],5:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="ibox float-e-margins">\r\n  <div class="ibox-content">\r\n    <div>\r\n      <span class="pull-right text-right">\r\n        <small>Average value of sales in the past month in: <strong>United states</strong></small>\r\n        <br/>\r\n        All sales: 162,862\r\n      </span>\r\n      <h3 class="font-bold no-margins">\r\n        Half-year revenue margin\r\n      </h3>\r\n      <small>Sales marketing.</small>\r\n    </div>\r\n\r\n    <div class="m-t-sm">\r\n\r\n      <div class="row">\r\n        <div class="col-md-8">\r\n          <div>\r\n            <canvas id="lineChart" height="114"></canvas>\r\n          </div>\r\n        </div>\r\n        <div class="col-md-4">\r\n          <ul class="stat-list m-t-lg">\r\n            <li>\r\n              <h2 class="no-margins">2,346</h2>\r\n              <small>Total orders in period</small>\r\n              <div class="progress progress-mini">\r\n                <div class="progress-bar" style="width: 48%;"></div>\r\n              </div>\r\n            </li>\r\n            <li>\r\n              <h2 class="no-margins ">4,422</h2>\r\n              <small>Orders in last month</small>\r\n              <div class="progress progress-mini">\r\n                <div class="progress-bar" style="width: 60%;"></div>\r\n              </div>\r\n            </li>\r\n          </ul>\r\n        </div>\r\n      </div>\r\n\r\n    </div>\r\n\r\n    <div class="m-t-md">\r\n      <small class="pull-right">\r\n        <i class="fa fa-clock-o"> </i>\r\n        Update on 16.07.2015\r\n      </small>\r\n      <small>\r\n        <strong>Analysis of sales:</strong> The value has been changed over time, and last month reached a level over $50,000.\r\n      </small>\r\n    </div>\r\n\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:30}],6:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="ibox float-e-margins">\r\n  <div class="ibox-title">\r\n    <h5>Monthly income</h5>\r\n    <div class="ibox-tools">\r\n      <span class="label label-primary">Updated 12.2015</span>\r\n    </div>\r\n  </div>\r\n  <div class="ibox-content no-padding">\r\n    <div class="flot-chart m-t-lg" style="height: 55px;">\r\n      <div class="flot-chart-content" id="flot-chart1"></div>\r\n    </div>\r\n  </div>\r\n\r\n</div>\r\n';return __p}},{underscore:30}],7:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="ibox float-e-margins">\r\n  <div class="ibox-title">\r\n    <h5>Custom responsive table </h5>\r\n    <div class="ibox-tools">\r\n      <a class="collapse-link">\r\n        <i class="fa fa-chevron-up"></i>\r\n      </a>\r\n      <a class="dropdown-toggle" data-toggle="dropdown" href="#">\r\n        <i class="fa fa-wrench"></i>\r\n      </a>\r\n      <ul class="dropdown-menu dropdown-user">\r\n        <li><a href="#">Config option 1</a>\r\n        </li>\r\n        <li><a href="#">Config option 2</a>\r\n        </li>\r\n      </ul>\r\n      <a class="close-link">\r\n        <i class="fa fa-times"></i>\r\n      </a>\r\n    </div>\r\n  </div>\r\n  <div class="ibox-content">\r\n    <div class="row">\r\n      <div class="col-sm-9 m-b-xs">\r\n        <div data-toggle="buttons" class="btn-group">\r\n          <label class="btn btn-sm btn-white"> <input type="radio" id="option1" name="options"> Day </label>\r\n          <label class="btn btn-sm btn-white active"> <input type="radio" id="option2" name="options"> Week </label>\r\n          <label class="btn btn-sm btn-white"> <input type="radio" id="option3" name="options"> Month </label>\r\n        </div>\r\n      </div>\r\n      <div class="col-sm-3">\r\n        <div class="input-group"><input type="text" placeholder="Search" class="input-sm form-control"> <span class="input-group-btn">\r\n          <button type="button" class="btn btn-sm btn-primary"> Go!</button> </span></div>\r\n        </div>\r\n      </div>\r\n      <div class="table-responsive">\r\n        <table class="table table-striped">\r\n          <thead>\r\n            <tr>\r\n\r\n              <th>#</th>\r\n              <th>Project </th>\r\n              <th>Name </th>\r\n              <th>Phone </th>\r\n              <th>Company </th>\r\n              <th>Completed </th>\r\n              <th>Task</th>\r\n              <th>Date</th>\r\n              <th>Action</th>\r\n            </tr>\r\n          </thead>\r\n          <tbody>\r\n            <tr>\r\n              <td>1</td>\r\n              <td>Project <small>This is example of project</small></td>\r\n              <td>Patrick Smith</td>\r\n              <td>0800 051213</td>\r\n              <td>Inceptos Hymenaeos Ltd</td>\r\n              <td><span class="pie">0.52/1.561</span></td>\r\n              <td>20%</td>\r\n              <td>Jul 14, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>2</td>\r\n              <td>Alpha project</td>\r\n              <td>Alice Jackson</td>\r\n              <td>0500 780909</td>\r\n              <td>Nec Euismod In Company</td>\r\n              <td><span class="pie">6,9</span></td>\r\n              <td>40%</td>\r\n              <td>Jul 16, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>3</td>\r\n              <td>Betha project</td>\r\n              <td>John Smith</td>\r\n              <td>0800 1111</td>\r\n              <td>Erat Volutpat</td>\r\n              <td><span class="pie">3,1</span></td>\r\n              <td>75%</td>\r\n              <td>Jul 18, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>4</td>\r\n              <td>Gamma project</td>\r\n              <td>Anna Jordan</td>\r\n              <td>(016977) 0648</td>\r\n              <td>Tellus Ltd</td>\r\n              <td><span class="pie">4,9</span></td>\r\n              <td>18%</td>\r\n              <td>Jul 22, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>2</td>\r\n              <td>Alpha project</td>\r\n              <td>Alice Jackson</td>\r\n              <td>0500 780909</td>\r\n              <td>Nec Euismod In Company</td>\r\n              <td><span class="pie">6,9</span></td>\r\n              <td>40%</td>\r\n              <td>Jul 16, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>1</td>\r\n              <td>Project <small>This is example of project</small></td>\r\n              <td>Patrick Smith</td>\r\n              <td>0800 051213</td>\r\n              <td>Inceptos Hymenaeos Ltd</td>\r\n              <td><span class="pie">0.52/1.561</span></td>\r\n              <td>20%</td>\r\n              <td>Jul 14, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>4</td>\r\n              <td>Gamma project</td>\r\n              <td>Anna Jordan</td>\r\n              <td>(016977) 0648</td>\r\n              <td>Tellus Ltd</td>\r\n              <td><span class="pie">4,9</span></td>\r\n              <td>18%</td>\r\n              <td>Jul 22, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>1</td>\r\n              <td>Project <small>This is example of project</small></td>\r\n              <td>Patrick Smith</td>\r\n              <td>0800 051213</td>\r\n              <td>Inceptos Hymenaeos Ltd</td>\r\n              <td><span class="pie">0.52/1.561</span></td>\r\n              <td>20%</td>\r\n              <td>Jul 14, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>2</td>\r\n              <td>Alpha project</td>\r\n              <td>Alice Jackson</td>\r\n              <td>0500 780909</td>\r\n              <td>Nec Euismod In Company</td>\r\n              <td><span class="pie">6,9</span></td>\r\n              <td>40%</td>\r\n              <td>Jul 16, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>3</td>\r\n              <td>Betha project</td>\r\n              <td>John Smith</td>\r\n              <td>0800 1111</td>\r\n              <td>Erat Volutpat</td>\r\n              <td><span class="pie">3,1</span></td>\r\n              <td>75%</td>\r\n              <td>Jul 18, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>4</td>\r\n              <td>Gamma project</td>\r\n              <td>Anna Jordan</td>\r\n              <td>(016977) 0648</td>\r\n              <td>Tellus Ltd</td>\r\n              <td><span class="pie">4,9</span></td>\r\n              <td>18%</td>\r\n              <td>Jul 22, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>2</td>\r\n              <td>Alpha project</td>\r\n              <td>Alice Jackson</td>\r\n              <td>0500 780909</td>\r\n              <td>Nec Euismod In Company</td>\r\n              <td><span class="pie">6,9</span></td>\r\n              <td>40%</td>\r\n              <td>Jul 16, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>1</td>\r\n              <td>Project <small>This is example of project</small></td>\r\n              <td>Patrick Smith</td>\r\n              <td>0800 051213</td>\r\n              <td>Inceptos Hymenaeos Ltd</td>\r\n              <td><span class="pie">0.52/1.561</span></td>\r\n              <td>20%</td>\r\n              <td>Jul 14, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n            <tr>\r\n              <td>4</td>\r\n              <td>Gamma project</td>\r\n              <td>Anna Jordan</td>\r\n              <td>(016977) 0648</td>\r\n              <td>Tellus Ltd</td>\r\n              <td><span class="pie">4,9</span></td>\r\n              <td>18%</td>\r\n              <td>Jul 22, 2013</td>\r\n              <td><a href="#"><i class="fa fa-check text-navy"></i></a></td>\r\n            </tr>\r\n          </tbody>\r\n        </table>\r\n      </div>\r\n    </div>\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:30}],8:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>Documents</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content">\r\n  <div class="row">\r\n    <div class="col-lg-3">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-content">\r\n          <div class="file-manager">\r\n            <h5>Show:</h5>\r\n            <a href="#" class="file-control active">Ale</a>\r\n            <a href="#" class="file-control">Documents</a>\r\n            <a href="#" class="file-control">Audio</a>\r\n            <a href="#" class="file-control">Images</a>\r\n            <div class="hr-line-dashed"></div>\r\n            <a href="#file-upload">\r\n              <button class="btn btn-primary btn-block">Upload Files</button>\r\n            </a>\r\n            <div class="hr-line-dashed"></div>\r\n            <h5>Folders</h5>\r\n            <ul class="folder-list" style="padding: 0">\r\n              <li><a href=""><i class="fa fa-folder"></i> Files</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Pictures</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Web pages</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Illustrations</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Films</a></li>\r\n              <li><a href=""><i class="fa fa-folder"></i> Books</a></li>\r\n            </ul>\r\n            <h5 class="tag-title">Tags</h5>\r\n            <ul class="tag-list" style="padding: 0">\r\n              <li><a href="">Family</a></li>\r\n              <li><a href="">Work</a></li>\r\n              <li><a href="">Home</a></li>\r\n              <li><a href="">Children</a></li>\r\n              <li><a href="">Holidays</a></li>\r\n              <li><a href="">Music</a></li>\r\n              <li><a href="">Photography</a></li>\r\n              <li><a href="">Film</a></li>\r\n            </ul>\r\n            <div class="clearfix"></div>\r\n          </div>\r\n        </div>\r\n      </div>\r\n    </div>\r\n    <div class="col-lg-9 animated fadeInRight">\r\n      <div class="row">\r\n        <div class="col-lg-12">\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-file"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Jan 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p1.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Italy street.jpg\r\n                  <br/>\r\n                  <small>Added: Jan 6, 2014</small>\r\n                </div>\r\n              </a>\r\n\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p2.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  My feel.png\r\n                  <br/>\r\n                  <small>Added: Jan 7, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-music"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Michal Jackson.mp3\r\n                  <br/>\r\n                  <small>Added: Jan 22, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="image">\r\n                  <img alt="image" class="img-responsive" src="/img/portal/p3.jpg">\r\n                </div>\r\n                <div class="file-name">\r\n                  Document_2014.doc\r\n                  <br/>\r\n                  <small>Added: Fab 11, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <div class="file">\r\n              <a href="#">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="img-responsive fa fa-film"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Monica\'s birthday.mpg4\r\n                  <br/>\r\n                  <small>Added: Fab 18, 2014</small>\r\n                </div>\r\n              </a>\r\n            </div>\r\n          </div>\r\n          <div class="file-box">\r\n            <a href="#">\r\n              <div class="file">\r\n                <span class="corner"></span>\r\n\r\n                <div class="icon">\r\n                  <i class="fa fa-bar-chart-o"></i>\r\n                </div>\r\n                <div class="file-name">\r\n                  Annual report 2014.xls\r\n                  <br/>\r\n                  <small>Added: Fab 22, 2014</small>\r\n                </div>\r\n              </div>\r\n            </a>\r\n          </div>\r\n\r\n        </div>\r\n      </div>\r\n    </div>\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:30}],9:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row wrapper border-bottom white-bg page-heading">\r\n  <div class="col-lg-9">\r\n    <h2>File Upload</h2>\r\n  </div>\r\n</div>\r\n<div class="wrapper wrapper-content animated fadeIn">\r\n  <div class="row">\r\n    <div class="col-lg-12">\r\n      <div class="ibox float-e-margins">\r\n        <div class="ibox-title">\r\n          <h5>Dropzone Area</h5>\r\n          <div class="ibox-tools">\r\n            <a class="collapse-link">\r\n              <i class="fa fa-chevron-up"></i>\r\n            </a>\r\n            <a class="dropdown-toggle" data-toggle="dropdown" href="#">\r\n              <i class="fa fa-wrench"></i>\r\n            </a>\r\n            <ul class="dropdown-menu dropdown-user">\r\n              <li><a href="#">Config option 1</a>\r\n              </li>\r\n              <li><a href="#">Config option 2</a>\r\n              </li>\r\n            </ul>\r\n            <a class="close-link">\r\n              <i class="fa fa-times"></i>\r\n            </a>\r\n          </div>\r\n        </div>\r\n        <div class="ibox-content">\r\n          <form id="my-awesome-dropzone" class="dropzone" action="#">\r\n            <div class="dropzone-previews"></div>\r\n            <button type="submit" class="btn btn-primary pull-right">Submit this form!</button>\r\n          </form>\r\n        </div>\r\n      </div>\r\n    </div>\r\n  </div>\r\n</div>\r\n';
+return __p}},{underscore:30}],10:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="footer">\r\n  <div class="pull-right">\r\n    10GB of <strong>250GB</strong> Free.\r\n  </div>\r\n  <div>\r\n    <strong>Copyright</strong> Example Company &copy; 2014-2015\r\n  </div>\r\n</div>\r\n';return __p}},{underscore:30}],11:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __t,__p="";Array.prototype.join}with(obj||{})__p+='<nav class="navbar-default navbar-static-side" role="navigation">\n    <div class="sidebar-collapse">\n        <ul class="nav" id="side-menu">\n            <li class="nav-header">\n                <div class="dropdown profile-element">\n                    <span>\n                        <img alt="image" class="img-circle" src="/img/portal/profile_small.jpg" />\n                    </span>\n                    <a data-toggle="dropdown" class="dropdown-toggle" href="#">\n                        <span class="clear">\n                          <span class="block m-t-xs">\n                            <strong class="font-bold">'+(null==(__t=name)?"":__t)+'</strong><b class="caret"></b>\n                          </span>\n                        </span>\n                    </a>\n                    <ul class="dropdown-menu animated fadeInRight m-t-xs">\n                        <li><a href="profile.html">Profile</a></li>\n                        <li><a href="contacts.html">Contacts</a></li>\n                        <li><a href="mailbox.html">Mailbox</a></li>\n                        <li class="divider"></li>\n                        <li><a href="login.html">Logout</a></li>\n                    </ul>\n                </div>\n            </li>\n            <li id="nav-dashboard">\n                <a href="#"><i class="fa fa-bar-chart-o"></i> <span class="nav-label">Dashboard</span></a>\n            </li>\n            <li id="nav-file-manager">\n                <a href="#file-manager"><i class="fa fa-files-o"></i> <span class="nav-label">Documents</span></a>\n            </li>\n        </ul>\n\n    </div>\n</nav>\n';return __p}},{underscore:30}],12:[function(require,module,exports){require("underscore");module.exports=function(obj){{var __p="";Array.prototype.join}with(obj||{})__p+='<div class="row border-bottom">\r\n  <nav class="navbar navbar-static-top" role="navigation" style="margin-bottom: 0">\r\n    <div class="navbar-header">\r\n      <a class="navbar-minimalize minimalize-styl-2 btn btn-primary " href="#"><i class="fa fa-bars"></i> </a>\r\n    </div>\r\n    <ul class="nav navbar-top-links navbar-right">\r\n      <li>\r\n        <span class="m-r-sm text-muted welcome-message">Welcome to INSPINIA+ Admin Theme.</span>\r\n      </li>\r\n      <li class="dropdown">\r\n        <a class="dropdown-toggle count-info" data-toggle="dropdown" href="#">\r\n          <i class="fa fa-envelope"></i>  <span class="label label-warning">16</span>\r\n        </a>\r\n        <ul class="dropdown-menu dropdown-messages">\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/a7.jpg">\r\n              </a>\r\n              <div class="media-body">\r\n                <small class="pull-right">46h ago</small>\r\n                <strong>Mike Loreipsum</strong> started following <strong>Monica Smith</strong>. <br>\r\n                <small class="text-muted">3 days ago at 7:58 pm - 10.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/a4.jpg">\r\n              </a>\r\n              <div class="media-body ">\r\n                <small class="pull-right text-navy">5h ago</small>\r\n                <strong>Chris Johnatan Overtunk</strong> started following <strong>Monica Smith</strong>. <br>\r\n                <small class="text-muted">Yesterday 1:21 pm - 11.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="dropdown-messages-box">\r\n              <a href="profile.html" class="pull-left">\r\n                <img alt="image" class="img-circle" src="/img/portal/profile.jpg">\r\n              </a>\r\n              <div class="media-body ">\r\n                <small class="pull-right">23h ago</small>\r\n                <strong>Monica Smith</strong> love <strong>Kim Smith</strong>. <br>\r\n                <small class="text-muted">2 days ago at 2:30 am - 11.06.2014</small>\r\n              </div>\r\n            </div>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="text-center link-block">\r\n              <a href="mailbox.html">\r\n                <i class="fa fa-envelope"></i> <strong>Read All Messages</strong>\r\n              </a>\r\n            </div>\r\n          </li>\r\n        </ul>\r\n      </li>\r\n      <li class="dropdown">\r\n        <a class="dropdown-toggle count-info" data-toggle="dropdown" href="#">\r\n          <i class="fa fa-bell"></i>  <span class="label label-primary">8</span>\r\n        </a>\r\n        <ul class="dropdown-menu dropdown-alerts">\r\n          <li>\r\n            <a href="mailbox.html">\r\n              <div>\r\n                <i class="fa fa-envelope fa-fw"></i> You have 16 messages\r\n                <span class="pull-right text-muted small">4 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <a href="profile.html">\r\n              <div>\r\n                <i class="fa fa-twitter fa-fw"></i> 3 New Followers\r\n                <span class="pull-right text-muted small">12 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <a href="grid_options.html">\r\n              <div>\r\n                <i class="fa fa-upload fa-fw"></i> Server Rebooted\r\n                <span class="pull-right text-muted small">4 minutes ago</span>\r\n              </div>\r\n            </a>\r\n          </li>\r\n          <li class="divider"></li>\r\n          <li>\r\n            <div class="text-center link-block">\r\n              <a href="notifications.html">\r\n                <strong>See All Alerts</strong>\r\n                <i class="fa fa-angle-right"></i>\r\n              </a>\r\n            </div>\r\n          </li>\r\n        </ul>\r\n      </li>\r\n\r\n\r\n      <li>\r\n        <a href="/login">\r\n          <i class="fa fa-sign-out"></i> Log out\r\n        </a>\r\n      </li>\r\n    </ul>\r\n\r\n  </nav>\r\n</div>\r\n';return __p}},{underscore:30}],13:[function(n,t){{var r=n("backbone"),e=n("backbone.broker");n("./file-upload")}t.exports=r.View.extend({initialize:function(){this.listenTo(e.channel("container"),"show",this.showView)},showView:function(n){this.setView(n)}})},{"./file-upload":19,backbone:28,"backbone.broker":24}],14:[function(n,t){var r=n("backbone"),e=n("./dashboard/table"),i=n("./dashboard/flot-chart"),a=n("./dashboard/chart-js");t.exports=r.View.extend({template:n("../templates/dashboard.jst"),onShow:function(){this.addSubView({view:new e,selector:"#table"}),this.addSubView({view:new i,selector:"#flot-chart"}),this.addSubView({view:new a,selector:"#chart-js"})}})},{"../templates/dashboard.jst":4,"./dashboard/chart-js":15,"./dashboard/flot-chart":16,"./dashboard/table":17,backbone:28}],15:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../../templates/dashboard/chart-js.jst"),onShow:function(){var n={labels:["January","February","March","April","May","June","July"],datasets:[{label:"Example dataset",fillColor:"rgba(220,220,220,0.5)",strokeColor:"rgba(220,220,220,1)",pointColor:"rgba(220,220,220,1)",pointStrokeColor:"#fff",pointHighlightFill:"#fff",pointHighlightStroke:"rgba(220,220,220,1)",data:[65,59,40,51,36,25,40]},{label:"Example dataset",fillColor:"rgba(26,179,148,0.5)",strokeColor:"rgba(26,179,148,0.7)",pointColor:"rgba(26,179,148,1)",pointStrokeColor:"#fff",pointHighlightFill:"#fff",pointHighlightStroke:"rgba(26,179,148,1)",data:[48,48,60,39,56,37,30]}]},t={scaleShowGridLines:!0,scaleGridLineColor:"rgba(0,0,0,.05)",scaleGridLineWidth:1,bezierCurve:!0,bezierCurveTension:.4,pointDot:!0,pointDotRadius:4,pointDotStrokeWidth:1,pointHitDetectionRadius:20,datasetStroke:!0,datasetStrokeWidth:2,datasetFill:!0,responsive:!0};if(this.$("#lineChart").length>0){var r=this.$("#lineChart")[0].getContext("2d");new Chart(r).Line(n,t)}}})},{"../../templates/dashboard/chart-js.jst":5,backbone:28}],16:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../../templates/dashboard/flot-chart.jst"),onShow:function(){var n=[[1262304e6,6],[12649824e5,3057],[12674016e5,20434],[127008e7,31982],[1272672e6,26602],[12753504e5,27826],[12779424e5,24302],[12806208e5,24237],[12832992e5,21004],[12858912e5,12144],[12885696e5,10577],[12911616e5,10295]],t=[[1262304e6,5],[12649824e5,200],[12674016e5,1605],[127008e7,6129],[1272672e6,11643],[12753504e5,19055],[12779424e5,30062],[12806208e5,39197],[12832992e5,37e3],[12858912e5,27e3],[12885696e5,21e3],[12911616e5,17e3]],r=[{label:"Data 1",data:n,color:"#17a084"},{label:"Data 2",data:t,color:"#127e68"}];$("#flot-chart1").length>0&&$.plot($("#flot-chart1"),r,{xaxis:{tickDecimals:0},series:{lines:{show:!0,fill:!0,fillColor:{colors:[{opacity:1},{opacity:1}]}},points:{width:.1,show:!1}},grid:{show:!1,borderWidth:0},legend:{show:!1}})}})},{"../../templates/dashboard/flot-chart.jst":6,backbone:28}],17:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../../templates/dashboard/table.jst"),events:{"click .collapse-link":"collapseIBox","click .close-link":"close"},collapseIBox:function(){var n=this.$(".collapse-link"),t=n.closest("div.ibox"),r=n.find("i"),e=t.find("div.ibox-content");e.slideToggle(200),r.toggleClass("fa-chevron-up").toggleClass("fa-chevron-down"),t.toggleClass("").toggleClass("border-bottom"),setTimeout(function(){t.resize(),t.find("[id^=map-]").resize()},50)},onShow:function(){this.$("span.pie").peity("pie",{fill:["#1ab394","#d7d7d7","#ffffff"]}),this.$(".line").peity("line",{fill:"#1ab394",stroke:"#169c81"}),this.$(".bar").peity("bar",{fill:["#1ab394","#d7d7d7"]}),this.$(".bar_dashboard").peity("bar",{fill:["#1ab394","#d7d7d7"],width:100});var n=this.$(".updating-chart").peity("line",{fill:"#1ab394",stroke:"#169c81",width:64}),t=Math.round(10*Math.random()),r=n.text().split(",");r.shift(),r.push(t),n.text(r.join(",")).change()}})},{"../../templates/dashboard/table.jst":7,backbone:28}],18:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/file-manager.jst")})},{"../templates/file-manager.jst":8,backbone:28}],19:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/file-upload.jst")})},{"../templates/file-upload.jst":9,backbone:28}],20:[function(n,t){var r=n("backbone");t.exports=r.View.extend({template:n("../templates/footer.jst")})},{"../templates/footer.jst":10,backbone:28}],21:[function(n,t){var r=n("backbone"),e=n("./nav"),i=n("./top-nav"),a=n("./container"),s=n("./footer"),l=n("../models/user");t.exports=r.View.extend({initialize:function(){this.addSubView({view:new e({model:new l}),selector:"#wrapper",location:"prepend"}),this.addSubView({view:new i,selector:"#page-wrapper"}),this.addSubView({view:new a,selector:"#page-wrapper"}),this.addSubView({view:new s,selector:"#page-wrapper"})}})},{"../models/user":2,"./container":13,"./footer":20,"./nav":22,"./top-nav":23,backbone:28}],22:[function(n,t){var r=n("backbone"),e=n("backbone.broker");t.exports=r.View.extend({template:n("../templates/nav.jst"),initialize:function(){$(window).bind("load resize",function(){$(this).width()<769?$("body").addClass("body-small"):$("body").removeClass("body-small")}),this.listenTo(e.channel("nav"),"select",this.highlightSelection),this.listenTo(e.channel("nav"),"toggle",this.toggleNav),this.listenTo(this.model,"change",this.render)},serializeData:function(){return this.model.toJSON()},onShow:function(){$("body").hasClass("fixed-sidebar")&&this.$(".sidebar-collapse").slimScroll({height:"100%",railOpacity:.9})},highlightSelection:function(n){this.$("li").removeClass("active"),this.$("#nav-"+n).addClass("active")},toggleNav:function(){$("body").toggleClass("mini-navbar");var n=function(){this.$("#side-menu").fadeIn(500)}.bind(this);!$("body").hasClass("mini-navbar")||$("body").hasClass("body-small")?(this.$("#side-menu").hide(),setTimeout(n,100)):$("body").hasClass("fixed-sidebar")?(this.$("#side-menu").hide(),setTimeout(n,300)):this.$("#side-menu").removeAttr("style")},onClose:function(){$(window).off("load resize")}})},{"../templates/nav.jst":11,backbone:28,"backbone.broker":24}],23:[function(n,t){var r=n("backbone"),e=n("backbone.broker");t.exports=r.View.extend({template:n("../templates/top-nav.jst"),events:{"click .navbar-minimalize":"toggleNavBar"},toggleNavBar:function(n){n.preventDefault(),e.channel("nav").publish("toggle")}})},{"../templates/top-nav.jst":12,backbone:28,"backbone.broker":24}],24:[function(n,t,r){!function(t){"function"==typeof define&&define.amd?define(["underscore","backbone","exports"],t):"object"==typeof r?t(n("underscore"),n("backbone"),r):Backbone.broker||(Backbone.broker={},t(_,Backbone,Backbone.broker))}(function(n,t,r){var e=!1,i=[],a=function(n){this.name=n,this.queue=[]};n.extend(a.prototype,t.Events,{publish:function(){return arguments.length>0&&this.queue.push(arguments),e&&(this.queue.forEach(function(n){this.trigger.apply(this,n)}.bind(this)),this.queue=[]),this},subscribe:function(){return this.on.apply(this,arguments),this},unsubscribe:function(){return this.off.apply(this,arguments),this}}),n.extend(r,{channel:function(n){if(n&&n.length>0){var t=i.filter(function(t){return t.name===n?t:void 0});if(t.length>0)return t[0];var r=new a(n);return i.push(r),r}},getChannels:function(){return i},start:function(){e=!0,i.forEach(function(n){n.publish()})},stop:function(){e=!1}})})},{backbone:28,underscore:25}],25:[function(n,t,r){(function(){var n=this,e=n._,i=Array.prototype,a=Object.prototype,s=Function.prototype,l=i.push,o=i.slice,c=i.concat,u=a.toString,d=a.hasOwnProperty,f=Array.isArray,h=Object.keys,p=s.bind,v=function(n){return n instanceof v?n:this instanceof v?void(this._wrapped=n):new v(n)};"undefined"!=typeof r?("undefined"!=typeof t&&t.exports&&(r=t.exports=v),r._=v):n._=v,v.VERSION="1.7.0";var m=function(n,t,r){if(void 0===t)return n;switch(null==r?3:r){case 1:return function(r){return n.call(t,r)};case 2:return function(r,e){return n.call(t,r,e)};case 3:return function(r,e,i){return n.call(t,r,e,i)};case 4:return function(r,e,i,a){return n.call(t,r,e,i,a)}}return function(){return n.apply(t,arguments)}};v.iteratee=function(n,t,r){return null==n?v.identity:v.isFunction(n)?m(n,t,r):v.isObject(n)?v.matches(n):v.property(n)},v.each=v.forEach=function(n,t,r){if(null==n)return n;t=m(t,r);var e,i=n.length;if(i===+i)for(e=0;i>e;e++)t(n[e],e,n);else{var a=v.keys(n);for(e=0,i=a.length;i>e;e++)t(n[a[e]],a[e],n)}return n},v.map=v.collect=function(n,t,r){if(null==n)return[];t=v.iteratee(t,r);for(var e,i=n.length!==+n.length&&v.keys(n),a=(i||n).length,s=Array(a),l=0;a>l;l++)e=i?i[l]:l,s[l]=t(n[e],e,n);return s};var g="Reduce of empty array with no initial value";v.reduce=v.foldl=v.inject=function(n,t,r,e){null==n&&(n=[]),t=m(t,e,4);var i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length,l=0;if(arguments.length<3){if(!s)throw new TypeError(g);r=n[a?a[l++]:l++]}for(;s>l;l++)i=a?a[l]:l,r=t(r,n[i],i,n);return r},v.reduceRight=v.foldr=function(n,t,r,e){null==n&&(n=[]),t=m(t,e,4);var i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;if(arguments.length<3){if(!s)throw new TypeError(g);r=n[a?a[--s]:--s]}for(;s--;)i=a?a[s]:s,r=t(r,n[i],i,n);return r},v.find=v.detect=function(n,t,r){var e;return t=v.iteratee(t,r),v.some(n,function(n,r,i){return t(n,r,i)?(e=n,!0):void 0}),e},v.filter=v.select=function(n,t,r){var e=[];return null==n?e:(t=v.iteratee(t,r),v.each(n,function(n,r,i){t(n,r,i)&&e.push(n)}),e)},v.reject=function(n,t,r){return v.filter(n,v.negate(v.iteratee(t)),r)},v.every=v.all=function(n,t,r){if(null==n)return!0;t=v.iteratee(t,r);var e,i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;for(e=0;s>e;e++)if(i=a?a[e]:e,!t(n[i],i,n))return!1;return!0},v.some=v.any=function(n,t,r){if(null==n)return!1;t=v.iteratee(t,r);var e,i,a=n.length!==+n.length&&v.keys(n),s=(a||n).length;for(e=0;s>e;e++)if(i=a?a[e]:e,t(n[i],i,n))return!0;return!1},v.contains=v.include=function(n,t){return null==n?!1:(n.length!==+n.length&&(n=v.values(n)),v.indexOf(n,t)>=0)},v.invoke=function(n,t){var r=o.call(arguments,2),e=v.isFunction(t);return v.map(n,function(n){return(e?t:n[t]).apply(n,r)})},v.pluck=function(n,t){return v.map(n,v.property(t))},v.where=function(n,t){return v.filter(n,v.matches(t))},v.findWhere=function(n,t){return v.find(n,v.matches(t))},v.max=function(n,t,r){var e,i,a=-1/0,s=-1/0;if(null==t&&null!=n){n=n.length===+n.length?n:v.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],e>a&&(a=e)}else t=v.iteratee(t,r),v.each(n,function(n,r,e){i=t(n,r,e),(i>s||i===-1/0&&a===-1/0)&&(a=n,s=i)});return a},v.min=function(n,t,r){var e,i,a=1/0,s=1/0;if(null==t&&null!=n){n=n.length===+n.length?n:v.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],a>e&&(a=e)}else t=v.iteratee(t,r),v.each(n,function(n,r,e){i=t(n,r,e),(s>i||1/0===i&&1/0===a)&&(a=n,s=i)});return a},v.shuffle=function(n){for(var t,r=n&&n.length===+n.length?n:v.values(n),e=r.length,i=Array(e),a=0;e>a;a++)t=v.random(0,a),t!==a&&(i[a]=i[t]),i[t]=r[a];return i},v.sample=function(n,t,r){return null==t||r?(n.length!==+n.length&&(n=v.values(n)),n[v.random(n.length-1)]):v.shuffle(n).slice(0,Math.max(0,t))},v.sortBy=function(n,t,r){return t=v.iteratee(t,r),v.pluck(v.map(n,function(n,r,e){return{value:n,index:r,criteria:t(n,r,e)}}).sort(function(n,t){var r=n.criteria,e=t.criteria;if(r!==e){if(r>e||void 0===r)return 1;if(e>r||void 0===e)return-1}return n.index-t.index}),"value")};var b=function(n){return function(t,r,e){var i={};return r=v.iteratee(r,e),v.each(t,function(e,a){var s=r(e,a,t);n(i,e,s)}),i}};v.groupBy=b(function(n,t,r){v.has(n,r)?n[r].push(t):n[r]=[t]}),v.indexBy=b(function(n,t,r){n[r]=t}),v.countBy=b(function(n,t,r){v.has(n,r)?n[r]++:n[r]=1}),v.sortedIndex=function(n,t,r,e){r=v.iteratee(r,e,1);for(var i=r(t),a=0,s=n.length;s>a;){var l=a+s>>>1;r(n[l])<i?a=l+1:s=l}return a},v.toArray=function(n){return n?v.isArray(n)?o.call(n):n.length===+n.length?v.map(n,v.identity):v.values(n):[]},v.size=function(n){return null==n?0:n.length===+n.length?n.length:v.keys(n).length},v.partition=function(n,t,r){t=v.iteratee(t,r);var e=[],i=[];return v.each(n,function(n,r,a){(t(n,r,a)?e:i).push(n)}),[e,i]},v.first=v.head=v.take=function(n,t,r){return null==n?void 0:null==t||r?n[0]:0>t?[]:o.call(n,0,t)},v.initial=function(n,t,r){return o.call(n,0,Math.max(0,n.length-(null==t||r?1:t)))},v.last=function(n,t,r){return null==n?void 0:null==t||r?n[n.length-1]:o.call(n,Math.max(n.length-t,0))},v.rest=v.tail=v.drop=function(n,t,r){return o.call(n,null==t||r?1:t)},v.compact=function(n){return v.filter(n,v.identity)};var y=function(n,t,r,e){if(t&&v.every(n,v.isArray))return c.apply(e,n);for(var i=0,a=n.length;a>i;i++){var s=n[i];v.isArray(s)||v.isArguments(s)?t?l.apply(e,s):y(s,t,r,e):r||e.push(s)}return e};v.flatten=function(n,t){return y(n,t,!1,[])},v.without=function(n){return v.difference(n,o.call(arguments,1))},v.uniq=v.unique=function(n,t,r,e){if(null==n)return[];v.isBoolean(t)||(e=r,r=t,t=!1),null!=r&&(r=v.iteratee(r,e));for(var i=[],a=[],s=0,l=n.length;l>s;s++){var o=n[s];if(t)s&&a===o||i.push(o),a=o;else if(r){var c=r(o,s,n);v.indexOf(a,c)<0&&(a.push(c),i.push(o))}else v.indexOf(i,o)<0&&i.push(o)}return i},v.union=function(){return v.uniq(y(arguments,!0,!0,[]))},v.intersection=function(n){if(null==n)return[];for(var t=[],r=arguments.length,e=0,i=n.length;i>e;e++){var a=n[e];if(!v.contains(t,a)){for(var s=1;r>s&&v.contains(arguments[s],a);s++);s===r&&t.push(a)}}return t},v.difference=function(n){var t=y(o.call(arguments,1),!0,!0,[]);return v.filter(n,function(n){return!v.contains(t,n)})},v.zip=function(n){if(null==n)return[];for(var t=v.max(arguments,"length").length,r=Array(t),e=0;t>e;e++)r[e]=v.pluck(arguments,e);return r},v.object=function(n,t){if(null==n)return{};for(var r={},e=0,i=n.length;i>e;e++)t?r[n[e]]=t[e]:r[n[e][0]]=n[e][1];return r},v.indexOf=function(n,t,r){if(null==n)return-1;var e=0,i=n.length;if(r){if("number"!=typeof r)return e=v.sortedIndex(n,t),n[e]===t?e:-1;e=0>r?Math.max(0,i+r):r}for(;i>e;e++)if(n[e]===t)return e;return-1},v.lastIndexOf=function(n,t,r){if(null==n)return-1;var e=n.length;for("number"==typeof r&&(e=0>r?e+r+1:Math.min(e,r+1));--e>=0;)if(n[e]===t)return e;return-1},v.range=function(n,t,r){arguments.length<=1&&(t=n||0,n=0),r=r||1;for(var e=Math.max(Math.ceil((t-n)/r),0),i=Array(e),a=0;e>a;a++,n+=r)i[a]=n;return i};var w=function(){};v.bind=function(n,t){var r,e;if(p&&n.bind===p)return p.apply(n,o.call(arguments,1));if(!v.isFunction(n))throw new TypeError("Bind must be called on a function");return r=o.call(arguments,2),e=function(){if(!(this instanceof e))return n.apply(t,r.concat(o.call(arguments)));w.prototype=n.prototype;var i=new w;w.prototype=null;var a=n.apply(i,r.concat(o.call(arguments)));return v.isObject(a)?a:i}},v.partial=function(n){var t=o.call(arguments,1);return function(){for(var r=0,e=t.slice(),i=0,a=e.length;a>i;i++)e[i]===v&&(e[i]=arguments[r++]);for(;r<arguments.length;)e.push(arguments[r++]);return n.apply(this,e)}},v.bindAll=function(n){var t,r,e=arguments.length;if(1>=e)throw new Error("bindAll must be passed function names");for(t=1;e>t;t++)r=arguments[t],n[r]=v.bind(n[r],n);return n},v.memoize=function(n,t){var r=function(e){var i=r.cache,a=t?t.apply(this,arguments):e;return v.has(i,a)||(i[a]=n.apply(this,arguments)),i[a]};return r.cache={},r},v.delay=function(n,t){var r=o.call(arguments,2);return setTimeout(function(){return n.apply(null,r)},t)},v.defer=function(n){return v.delay.apply(v,[n,1].concat(o.call(arguments,1)))},v.throttle=function(n,t,r){var e,i,a,s=null,l=0;r||(r={});var o=function(){l=r.leading===!1?0:v.now(),s=null,a=n.apply(e,i),s||(e=i=null)};return function(){var c=v.now();l||r.leading!==!1||(l=c);var u=t-(c-l);return e=this,i=arguments,0>=u||u>t?(clearTimeout(s),s=null,l=c,a=n.apply(e,i),s||(e=i=null)):s||r.trailing===!1||(s=setTimeout(o,u)),a}},v.debounce=function(n,t,r){var e,i,a,s,l,o=function(){var c=v.now()-s;t>c&&c>0?e=setTimeout(o,t-c):(e=null,r||(l=n.apply(a,i),e||(a=i=null)))};return function(){a=this,i=arguments,s=v.now();var c=r&&!e;return e||(e=setTimeout(o,t)),c&&(l=n.apply(a,i),a=i=null),l}},v.wrap=function(n,t){return v.partial(t,n)},v.negate=function(n){return function(){return!n.apply(this,arguments)}},v.compose=function(){var n=arguments,t=n.length-1;return function(){for(var r=t,e=n[t].apply(this,arguments);r--;)e=n[r].call(this,e);return e}},v.after=function(n,t){return function(){return--n<1?t.apply(this,arguments):void 0}},v.before=function(n,t){var r;return function(){return--n>0?r=t.apply(this,arguments):t=null,r}},v.once=v.partial(v.before,2),v.keys=function(n){if(!v.isObject(n))return[];if(h)return h(n);var t=[];for(var r in n)v.has(n,r)&&t.push(r);return t},v.values=function(n){for(var t=v.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=n[t[i]];return e},v.pairs=function(n){for(var t=v.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=[t[i],n[t[i]]];return e},v.invert=function(n){for(var t={},r=v.keys(n),e=0,i=r.length;i>e;e++)t[n[r[e]]]=r[e];return t},v.functions=v.methods=function(n){var t=[];for(var r in n)v.isFunction(n[r])&&t.push(r);return t.sort()},v.extend=function(n){if(!v.isObject(n))return n;for(var t,r,e=1,i=arguments.length;i>e;e++){t=arguments[e];for(r in t)d.call(t,r)&&(n[r]=t[r])}return n},v.pick=function(n,t,r){var e,i={};if(null==n)return i;if(v.isFunction(t)){t=m(t,r);for(e in n){var a=n[e];t(a,e,n)&&(i[e]=a)}}else{var s=c.apply([],o.call(arguments,1));n=new Object(n);for(var l=0,u=s.length;u>l;l++)e=s[l],e in n&&(i[e]=n[e])}return i},v.omit=function(n,t,r){if(v.isFunction(t))t=v.negate(t);else{var e=v.map(c.apply([],o.call(arguments,1)),String);t=function(n,t){return!v.contains(e,t)}}return v.pick(n,t,r)},v.defaults=function(n){if(!v.isObject(n))return n;for(var t=1,r=arguments.length;r>t;t++){var e=arguments[t];for(var i in e)void 0===n[i]&&(n[i]=e[i])}return n},v.clone=function(n){return v.isObject(n)?v.isArray(n)?n.slice():v.extend({},n):n},v.tap=function(n,t){return t(n),n};var x=function(n,t,r,e){if(n===t)return 0!==n||1/n===1/t;if(null==n||null==t)return n===t;n instanceof v&&(n=n._wrapped),t instanceof v&&(t=t._wrapped);var i=u.call(n);if(i!==u.call(t))return!1;switch(i){case"[object RegExp]":case"[object String]":return""+n==""+t;case"[object Number]":return+n!==+n?+t!==+t:0===+n?1/+n===1/t:+n===+t;case"[object Date]":case"[object Boolean]":return+n===+t}if("object"!=typeof n||"object"!=typeof t)return!1;for(var a=r.length;a--;)if(r[a]===n)return e[a]===t;var s=n.constructor,l=t.constructor;if(s!==l&&"constructor"in n&&"constructor"in t&&!(v.isFunction(s)&&s instanceof s&&v.isFunction(l)&&l instanceof l))return!1;r.push(n),e.push(t);var o,c;if("[object Array]"===i){if(o=n.length,c=o===t.length)for(;o--&&(c=x(n[o],t[o],r,e)););}else{var d,f=v.keys(n);if(o=f.length,c=v.keys(t).length===o)for(;o--&&(d=f[o],c=v.has(t,d)&&x(n[d],t[d],r,e)););}return r.pop(),e.pop(),c};v.isEqual=function(n,t){return x(n,t,[],[])},v.isEmpty=function(n){if(null==n)return!0;if(v.isArray(n)||v.isString(n)||v.isArguments(n))return 0===n.length;for(var t in n)if(v.has(n,t))return!1;return!0},v.isElement=function(n){return!(!n||1!==n.nodeType)},v.isArray=f||function(n){return"[object Array]"===u.call(n)},v.isObject=function(n){var t=typeof n;return"function"===t||"object"===t&&!!n},v.each(["Arguments","Function","String","Number","Date","RegExp"],function(n){v["is"+n]=function(t){return u.call(t)==="[object "+n+"]"}}),v.isArguments(arguments)||(v.isArguments=function(n){return v.has(n,"callee")}),"function"!=typeof/./&&(v.isFunction=function(n){return"function"==typeof n||!1}),v.isFinite=function(n){return isFinite(n)&&!isNaN(parseFloat(n))},v.isNaN=function(n){return v.isNumber(n)&&n!==+n},v.isBoolean=function(n){return n===!0||n===!1||"[object Boolean]"===u.call(n)},v.isNull=function(n){return null===n},v.isUndefined=function(n){return void 0===n},v.has=function(n,t){return null!=n&&d.call(n,t)},v.noConflict=function(){return n._=e,this},v.identity=function(n){return n},v.constant=function(n){return function(){return n}},v.noop=function(){},v.property=function(n){return function(t){return t[n]}},v.matches=function(n){var t=v.pairs(n),r=t.length;return function(n){if(null==n)return!r;n=new Object(n);for(var e=0;r>e;e++){var i=t[e],a=i[0];if(i[1]!==n[a]||!(a in n))return!1}return!0}},v.times=function(n,t,r){var e=Array(Math.max(0,n));t=m(t,r,1);for(var i=0;n>i;i++)e[i]=t(i);return e},v.random=function(n,t){return null==t&&(t=n,n=0),n+Math.floor(Math.random()*(t-n+1))},v.now=Date.now||function(){return(new Date).getTime()};var _={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#x27;","`":"&#x60;"},k=v.invert(_),j=function(n){var t=function(t){return n[t]},r="(?:"+v.keys(n).join("|")+")",e=RegExp(r),i=RegExp(r,"g");return function(n){return n=null==n?"":""+n,e.test(n)?n.replace(i,t):n}};v.escape=j(_),v.unescape=j(k),v.result=function(n,t){if(null==n)return void 0;var r=n[t];return v.isFunction(r)?n[t]():r};var A=0;v.uniqueId=function(n){var t=++A+"";return n?n+t:t},v.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g,escape:/<%-([\s\S]+?)%>/g};var S=/(.)^/,E={"'":"'","\\":"\\","\r":"r","\n":"n","\u2028":"u2028","\u2029":"u2029"},T=/\\|'|\r|\n|\u2028|\u2029/g,O=function(n){return"\\"+E[n]};v.template=function(n,t,r){!t&&r&&(t=r),t=v.defaults({},t,v.templateSettings);var e=RegExp([(t.escape||S).source,(t.interpolate||S).source,(t.evaluate||S).source].join("|")+"|$","g"),i=0,a="__p+='";n.replace(e,function(t,r,e,s,l){return a+=n.slice(i,l).replace(T,O),i=l+t.length,r?a+="'+\n((__t=("+r+"))==null?'':_.escape(__t))+\n'":e?a+="'+\n((__t=("+e+"))==null?'':__t)+\n'":s&&(a+="';\n"+s+"\n__p+='"),t}),a+="';\n",t.variable||(a="with(obj||{}){\n"+a+"}\n"),a="var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n"+a+"return __p;\n";try{var s=new Function(t.variable||"obj","_",a)}catch(l){throw l.source=a,l}var o=function(n){return s.call(this,n,v)},c=t.variable||"obj";return o.source="function("+c+"){\n"+a+"}",o},v.chain=function(n){var t=v(n);return t._chain=!0,t};var $=function(n){return this._chain?v(n).chain():n};v.mixin=function(n){v.each(v.functions(n),function(t){var r=v[t]=n[t];v.prototype[t]=function(){var n=[this._wrapped];return l.apply(n,arguments),$.call(this,r.apply(v,n))}})},v.mixin(v),v.each(["pop","push","reverse","shift","sort","splice","unshift"],function(n){var t=i[n];v.prototype[n]=function(){var r=this._wrapped;return t.apply(r,arguments),"shift"!==n&&"splice"!==n||0!==r.length||delete r[0],$.call(this,r)}}),v.each(["concat","join","slice"],function(n){var t=i[n];v.prototype[n]=function(){return $.call(this,t.apply(this._wrapped,arguments))}}),v.prototype.value=function(){return this._wrapped},"function"==typeof define&&define.amd&&define("underscore",[],function(){return v})}).call(this)},{}],26:[function(n,t,r){!function(t){"function"==typeof define&&define.amd?define(["underscore","backbone"],t):"object"==typeof r?t(n("underscore"),n("backbone")):t(_,Backbone)}(function(n,t){n.extend(t.View.prototype,{render:function(){if(this.template&&n.isFunction(this.template)){var t=n.isFunction(this.serializeData)?this.serializeData():this,r=$(this.template(t));this.attachToTemplate&&1===r.length?(this.$el.replaceWith(r),this.setElement(r)):this.$el.html(r)}return this.restoreSubViews(),this.onRender&&n.isFunction(this.onRender)&&this.onRender.apply(this,arguments),this.trigger("rendered",this),this},setView:function(n,t){return this.removeSubViews(),t&&t.emptyDOM&&this.$el.empty(),this.addSubView({view:n,selector:this.$el}),n},getView:function(){return this._subViews&&this._subViews.length>0?this._subViews[0]:void 0},addSubView:function(n){if(!n||!n.view)throw new Error("Missing required view option");return this._subViews?this._subViews.push(n):this._subViews=[n],this.listenTo(n.view,"closed",this._removeSubView),this._showSubView(n)},restoreSubViews:function(){n.each(this._subViews,this._showSubView,this)},_showSubView:function(t){var r;return r=n.isObject(t.selector)?t.selector:n.isString(t.selector)?this.$(t.selector):this.$el,t.view.render(),"prepend"===t.location?r.prepend(t.view.el):"before"===t.location?r.before(t.view.el):"after"===t.location?r.after(t.view.el):r.append(t.view.el),t.view.onShow&&n.isFunction(t.view.onShow)&&t.view.onShow.apply(t.view,arguments),t.view.trigger("shown",this),t.view},getSubViewCount:function(){return this._subViews?this._subViews.length:0},removeSubViews:function(){n.each(this._subViews,function(n){this.stopListening(n.view),n.view.close()},this),this._subViews=[]},_removeSubView:function(t){var r=n.findWhere(this._subViews,{view:t}),e=this._subViews.indexOf(r);e>-1&&this._subViews.splice(e,1)},removeSubViewForModel:function(t){n.find(this._subViews,function(n){return n.view.model&&n.view.model===t?(n.view.close(),!0):void 0},this)},getAttr:function(n){return this.model?this.model.get(n):void 0},serializeForm:function(t){var r={},e=this.$(t).serializeArray();return n.each(e,function(n){r[n.name]?(r[n.name].push||(r[n.name]=[r[n.name]]),r[n.name].push(n.value||"")):r[n.name]=n.value||""
+}),r},close:function(){this.onClose&&n.isFunction(this.onClose)&&this.onClose.apply(this,arguments),this.removeSubViews(),this.remove(),this.trigger("closed",this),this.unbind()}})})},{backbone:28,underscore:27}],27:[function(n,t,r){arguments[4][25][0].apply(r,arguments)},{dup:25}],28:[function(n,t,r){!function(t,e){if("function"==typeof define&&define.amd)define(["underscore","jquery","exports"],function(n,r,i){t.Backbone=e(t,i,n,r)});else if("undefined"!=typeof r){var i=n("underscore");e(t,r,i)}else t.Backbone=e(t,{},t._,t.jQuery||t.Zepto||t.ender||t.$)}(this,function(n,t,r,e){{var i=n.Backbone,a=[],s=(a.push,a.slice);a.splice}t.VERSION="1.1.2",t.$=e,t.noConflict=function(){return n.Backbone=i,this},t.emulateHTTP=!1,t.emulateJSON=!1;var l=t.Events={on:function(n,t,r){if(!c(this,"on",n,[t,r])||!t)return this;this._events||(this._events={});var e=this._events[n]||(this._events[n]=[]);return e.push({callback:t,context:r,ctx:r||this}),this},once:function(n,t,e){if(!c(this,"once",n,[t,e])||!t)return this;var i=this,a=r.once(function(){i.off(n,a),t.apply(this,arguments)});return a._callback=t,this.on(n,a,e)},off:function(n,t,e){var i,a,s,l,o,u,d,f;if(!this._events||!c(this,"off",n,[t,e]))return this;if(!n&&!t&&!e)return this._events=void 0,this;for(l=n?[n]:r.keys(this._events),o=0,u=l.length;u>o;o++)if(n=l[o],s=this._events[n]){if(this._events[n]=i=[],t||e)for(d=0,f=s.length;f>d;d++)a=s[d],(t&&t!==a.callback&&t!==a.callback._callback||e&&e!==a.context)&&i.push(a);i.length||delete this._events[n]}return this},trigger:function(n){if(!this._events)return this;var t=s.call(arguments,1);if(!c(this,"trigger",n,t))return this;var r=this._events[n],e=this._events.all;return r&&u(r,t),e&&u(e,arguments),this},stopListening:function(n,t,e){var i=this._listeningTo;if(!i)return this;var a=!t&&!e;e||"object"!=typeof t||(e=this),n&&((i={})[n._listenId]=n);for(var s in i)n=i[s],n.off(t,e,this),(a||r.isEmpty(n._events))&&delete this._listeningTo[s];return this}},o=/\s+/,c=function(n,t,r,e){if(!r)return!0;if("object"==typeof r){for(var i in r)n[t].apply(n,[i,r[i]].concat(e));return!1}if(o.test(r)){for(var a=r.split(o),s=0,l=a.length;l>s;s++)n[t].apply(n,[a[s]].concat(e));return!1}return!0},u=function(n,t){var r,e=-1,i=n.length,a=t[0],s=t[1],l=t[2];switch(t.length){case 0:for(;++e<i;)(r=n[e]).callback.call(r.ctx);return;case 1:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a);return;case 2:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a,s);return;case 3:for(;++e<i;)(r=n[e]).callback.call(r.ctx,a,s,l);return;default:for(;++e<i;)(r=n[e]).callback.apply(r.ctx,t);return}},d={listenTo:"on",listenToOnce:"once"};r.each(d,function(n,t){l[t]=function(t,e,i){var a=this._listeningTo||(this._listeningTo={}),s=t._listenId||(t._listenId=r.uniqueId("l"));return a[s]=t,i||"object"!=typeof e||(i=this),t[n](e,i,this),this}}),l.bind=l.on,l.unbind=l.off,r.extend(t,l);var f=t.Model=function(n,t){var e=n||{};t||(t={}),this.cid=r.uniqueId("c"),this.attributes={},t.collection&&(this.collection=t.collection),t.parse&&(e=this.parse(e,t)||{}),e=r.defaults({},e,r.result(this,"defaults")),this.set(e,t),this.changed={},this.initialize.apply(this,arguments)};r.extend(f.prototype,l,{changed:null,validationError:null,idAttribute:"id",initialize:function(){},toJSON:function(){return r.clone(this.attributes)},sync:function(){return t.sync.apply(this,arguments)},get:function(n){return this.attributes[n]},escape:function(n){return r.escape(this.get(n))},has:function(n){return null!=this.get(n)},set:function(n,t,e){var i,a,s,l,o,c,u,d;if(null==n)return this;if("object"==typeof n?(a=n,e=t):(a={})[n]=t,e||(e={}),!this._validate(a,e))return!1;s=e.unset,o=e.silent,l=[],c=this._changing,this._changing=!0,c||(this._previousAttributes=r.clone(this.attributes),this.changed={}),d=this.attributes,u=this._previousAttributes,this.idAttribute in a&&(this.id=a[this.idAttribute]);for(i in a)t=a[i],r.isEqual(d[i],t)||l.push(i),r.isEqual(u[i],t)?delete this.changed[i]:this.changed[i]=t,s?delete d[i]:d[i]=t;if(!o){l.length&&(this._pending=e);for(var f=0,h=l.length;h>f;f++)this.trigger("change:"+l[f],this,d[l[f]],e)}if(c)return this;if(!o)for(;this._pending;)e=this._pending,this._pending=!1,this.trigger("change",this,e);return this._pending=!1,this._changing=!1,this},unset:function(n,t){return this.set(n,void 0,r.extend({},t,{unset:!0}))},clear:function(n){var t={};for(var e in this.attributes)t[e]=void 0;return this.set(t,r.extend({},n,{unset:!0}))},hasChanged:function(n){return null==n?!r.isEmpty(this.changed):r.has(this.changed,n)},changedAttributes:function(n){if(!n)return this.hasChanged()?r.clone(this.changed):!1;var t,e=!1,i=this._changing?this._previousAttributes:this.attributes;for(var a in n)r.isEqual(i[a],t=n[a])||((e||(e={}))[a]=t);return e},previous:function(n){return null!=n&&this._previousAttributes?this._previousAttributes[n]:null},previousAttributes:function(){return r.clone(this._previousAttributes)},fetch:function(n){n=n?r.clone(n):{},void 0===n.parse&&(n.parse=!0);var t=this,e=n.success;return n.success=function(r){return t.set(t.parse(r,n),n)?(e&&e(t,r,n),void t.trigger("sync",t,r,n)):!1},J(this,n),this.sync("read",this,n)},save:function(n,t,e){var i,a,s,l=this.attributes;if(null==n||"object"==typeof n?(i=n,e=t):(i={})[n]=t,e=r.extend({validate:!0},e),i&&!e.wait){if(!this.set(i,e))return!1}else if(!this._validate(i,e))return!1;i&&e.wait&&(this.attributes=r.extend({},l,i)),void 0===e.parse&&(e.parse=!0);var o=this,c=e.success;return e.success=function(n){o.attributes=l;var t=o.parse(n,e);return e.wait&&(t=r.extend(i||{},t)),r.isObject(t)&&!o.set(t,e)?!1:(c&&c(o,n,e),void o.trigger("sync",o,n,e))},J(this,e),a=this.isNew()?"create":e.patch?"patch":"update","patch"===a&&(e.attrs=i),s=this.sync(a,this,e),i&&e.wait&&(this.attributes=l),s},destroy:function(n){n=n?r.clone(n):{};var t=this,e=n.success,i=function(){t.trigger("destroy",t,t.collection,n)};if(n.success=function(r){(n.wait||t.isNew())&&i(),e&&e(t,r,n),t.isNew()||t.trigger("sync",t,r,n)},this.isNew())return n.success(),!1;J(this,n);var a=this.sync("delete",this,n);return n.wait||i(),a},url:function(){var n=r.result(this,"urlRoot")||r.result(this.collection,"url")||M();return this.isNew()?n:n.replace(/([^\/])$/,"$1/")+encodeURIComponent(this.id)},parse:function(n){return n},clone:function(){return new this.constructor(this.attributes)},isNew:function(){return!this.has(this.idAttribute)},isValid:function(n){return this._validate({},r.extend(n||{},{validate:!0}))},_validate:function(n,t){if(!t.validate||!this.validate)return!0;n=r.extend({},this.attributes,n);var e=this.validationError=this.validate(n,t)||null;return e?(this.trigger("invalid",this,e,r.extend(t,{validationError:e})),!1):!0}});var h=["keys","values","pairs","invert","pick","omit"];r.each(h,function(n){f.prototype[n]=function(){var t=s.call(arguments);return t.unshift(this.attributes),r[n].apply(r,t)}});var p=t.Collection=function(n,t){t||(t={}),t.model&&(this.model=t.model),void 0!==t.comparator&&(this.comparator=t.comparator),this._reset(),this.initialize.apply(this,arguments),n&&this.reset(n,r.extend({silent:!0},t))},v={add:!0,remove:!0,merge:!0},m={add:!0,remove:!1};r.extend(p.prototype,l,{model:f,initialize:function(){},toJSON:function(n){return this.map(function(t){return t.toJSON(n)})},sync:function(){return t.sync.apply(this,arguments)},add:function(n,t){return this.set(n,r.extend({merge:!1},t,m))},remove:function(n,t){var e=!r.isArray(n);n=e?[n]:r.clone(n),t||(t={});var i,a,s,l;for(i=0,a=n.length;a>i;i++)l=n[i]=this.get(n[i]),l&&(delete this._byId[l.id],delete this._byId[l.cid],s=this.indexOf(l),this.models.splice(s,1),this.length--,t.silent||(t.index=s,l.trigger("remove",l,this,t)),this._removeReference(l,t));return e?n[0]:n},set:function(n,t){t=r.defaults({},t,v),t.parse&&(n=this.parse(n,t));var e=!r.isArray(n);n=e?n?[n]:[]:r.clone(n);var i,a,s,l,o,c,u,d=t.at,h=this.model,p=this.comparator&&null==d&&t.sort!==!1,m=r.isString(this.comparator)?this.comparator:null,g=[],b=[],y={},w=t.add,x=t.merge,_=t.remove,k=!p&&w&&_?[]:!1;for(i=0,a=n.length;a>i;i++){if(o=n[i]||{},s=o instanceof f?l=o:o[h.prototype.idAttribute||"id"],c=this.get(s))_&&(y[c.cid]=!0),x&&(o=o===l?l.attributes:o,t.parse&&(o=c.parse(o,t)),c.set(o,t),p&&!u&&c.hasChanged(m)&&(u=!0)),n[i]=c;else if(w){if(l=n[i]=this._prepareModel(o,t),!l)continue;g.push(l),this._addReference(l,t)}l=c||l,!k||!l.isNew()&&y[l.id]||k.push(l),y[l.id]=!0}if(_){for(i=0,a=this.length;a>i;++i)y[(l=this.models[i]).cid]||b.push(l);b.length&&this.remove(b,t)}if(g.length||k&&k.length)if(p&&(u=!0),this.length+=g.length,null!=d)for(i=0,a=g.length;a>i;i++)this.models.splice(d+i,0,g[i]);else{k&&(this.models.length=0);var j=k||g;for(i=0,a=j.length;a>i;i++)this.models.push(j[i])}if(u&&this.sort({silent:!0}),!t.silent){for(i=0,a=g.length;a>i;i++)(l=g[i]).trigger("add",l,this,t);(u||k&&k.length)&&this.trigger("sort",this,t)}return e?n[0]:n},reset:function(n,t){t||(t={});for(var e=0,i=this.models.length;i>e;e++)this._removeReference(this.models[e],t);return t.previousModels=this.models,this._reset(),n=this.add(n,r.extend({silent:!0},t)),t.silent||this.trigger("reset",this,t),n},push:function(n,t){return this.add(n,r.extend({at:this.length},t))},pop:function(n){var t=this.at(this.length-1);return this.remove(t,n),t},unshift:function(n,t){return this.add(n,r.extend({at:0},t))},shift:function(n){var t=this.at(0);return this.remove(t,n),t},slice:function(){return s.apply(this.models,arguments)},get:function(n){return null==n?void 0:this._byId[n]||this._byId[n.id]||this._byId[n.cid]},at:function(n){return this.models[n]},where:function(n,t){return r.isEmpty(n)?t?void 0:[]:this[t?"find":"filter"](function(t){for(var r in n)if(n[r]!==t.get(r))return!1;return!0})},findWhere:function(n){return this.where(n,!0)},sort:function(n){if(!this.comparator)throw new Error("Cannot sort a set without a comparator");return n||(n={}),r.isString(this.comparator)||1===this.comparator.length?this.models=this.sortBy(this.comparator,this):this.models.sort(r.bind(this.comparator,this)),n.silent||this.trigger("sort",this,n),this},pluck:function(n){return r.invoke(this.models,"get",n)},fetch:function(n){n=n?r.clone(n):{},void 0===n.parse&&(n.parse=!0);var t=n.success,e=this;return n.success=function(r){var i=n.reset?"reset":"set";e[i](r,n),t&&t(e,r,n),e.trigger("sync",e,r,n)},J(this,n),this.sync("read",this,n)},create:function(n,t){if(t=t?r.clone(t):{},!(n=this._prepareModel(n,t)))return!1;t.wait||this.add(n,t);var e=this,i=t.success;return t.success=function(n,r){t.wait&&e.add(n,t),i&&i(n,r,t)},n.save(null,t),n},parse:function(n){return n},clone:function(){return new this.constructor(this.models)},_reset:function(){this.length=0,this.models=[],this._byId={}},_prepareModel:function(n,t){if(n instanceof f)return n;t=t?r.clone(t):{},t.collection=this;var e=new this.model(n,t);return e.validationError?(this.trigger("invalid",this,e.validationError,t),!1):e},_addReference:function(n){this._byId[n.cid]=n,null!=n.id&&(this._byId[n.id]=n),n.collection||(n.collection=this),n.on("all",this._onModelEvent,this)},_removeReference:function(n){this===n.collection&&delete n.collection,n.off("all",this._onModelEvent,this)},_onModelEvent:function(n,t,r,e){("add"!==n&&"remove"!==n||r===this)&&("destroy"===n&&this.remove(t,e),t&&n==="change:"+t.idAttribute&&(delete this._byId[t.previous(t.idAttribute)],null!=t.id&&(this._byId[t.id]=t)),this.trigger.apply(this,arguments))}});var g=["forEach","each","map","collect","reduce","foldl","inject","reduceRight","foldr","find","detect","filter","select","reject","every","all","some","any","include","contains","invoke","max","min","toArray","size","first","head","take","initial","rest","tail","drop","last","without","difference","indexOf","shuffle","lastIndexOf","isEmpty","chain","sample"];r.each(g,function(n){p.prototype[n]=function(){var t=s.call(arguments);return t.unshift(this.models),r[n].apply(r,t)}});var b=["groupBy","countBy","sortBy","indexBy"];r.each(b,function(n){p.prototype[n]=function(t,e){var i=r.isFunction(t)?t:function(n){return n.get(t)};return r[n](this.models,i,e)}});var y=t.View=function(n){this.cid=r.uniqueId("view"),n||(n={}),r.extend(this,r.pick(n,x)),this._ensureElement(),this.initialize.apply(this,arguments),this.delegateEvents()},w=/^(\S+)\s*(.*)$/,x=["model","collection","el","id","attributes","className","tagName","events"];r.extend(y.prototype,l,{tagName:"div",$:function(n){return this.$el.find(n)},initialize:function(){},render:function(){return this},remove:function(){return this.$el.remove(),this.stopListening(),this},setElement:function(n,r){return this.$el&&this.undelegateEvents(),this.$el=n instanceof t.$?n:t.$(n),this.el=this.$el[0],r!==!1&&this.delegateEvents(),this},delegateEvents:function(n){if(!n&&!(n=r.result(this,"events")))return this;this.undelegateEvents();for(var t in n){var e=n[t];if(r.isFunction(e)||(e=this[n[t]]),e){var i=t.match(w),a=i[1],s=i[2];e=r.bind(e,this),a+=".delegateEvents"+this.cid,""===s?this.$el.on(a,e):this.$el.on(a,s,e)}}return this},undelegateEvents:function(){return this.$el.off(".delegateEvents"+this.cid),this},_ensureElement:function(){if(this.el)this.setElement(r.result(this,"el"),!1);else{var n=r.extend({},r.result(this,"attributes"));this.id&&(n.id=r.result(this,"id")),this.className&&(n["class"]=r.result(this,"className"));var e=t.$("<"+r.result(this,"tagName")+">").attr(n);this.setElement(e,!1)}}}),t.sync=function(n,e,i){var a=k[n];r.defaults(i||(i={}),{emulateHTTP:t.emulateHTTP,emulateJSON:t.emulateJSON});var s={type:a,dataType:"json"};if(i.url||(s.url=r.result(e,"url")||M()),null!=i.data||!e||"create"!==n&&"update"!==n&&"patch"!==n||(s.contentType="application/json",s.data=JSON.stringify(i.attrs||e.toJSON(i))),i.emulateJSON&&(s.contentType="application/x-www-form-urlencoded",s.data=s.data?{model:s.data}:{}),i.emulateHTTP&&("PUT"===a||"DELETE"===a||"PATCH"===a)){s.type="POST",i.emulateJSON&&(s.data._method=a);var l=i.beforeSend;i.beforeSend=function(n){return n.setRequestHeader("X-HTTP-Method-Override",a),l?l.apply(this,arguments):void 0}}"GET"===s.type||i.emulateJSON||(s.processData=!1),"PATCH"===s.type&&_&&(s.xhr=function(){return new ActiveXObject("Microsoft.XMLHTTP")});var o=i.xhr=t.ajax(r.extend(s,i));return e.trigger("request",e,o,i),o};var _=!("undefined"==typeof window||!window.ActiveXObject||window.XMLHttpRequest&&(new XMLHttpRequest).dispatchEvent),k={create:"POST",update:"PUT",patch:"PATCH","delete":"DELETE",read:"GET"};t.ajax=function(){return t.$.ajax.apply(t.$,arguments)};var j=t.Router=function(n){n||(n={}),n.routes&&(this.routes=n.routes),this._bindRoutes(),this.initialize.apply(this,arguments)},A=/\((.*?)\)/g,S=/(\(\?)?:\w+/g,E=/\*\w+/g,T=/[\-{}\[\]+?.,\\\^$|#\s]/g;r.extend(j.prototype,l,{initialize:function(){},route:function(n,e,i){r.isRegExp(n)||(n=this._routeToRegExp(n)),r.isFunction(e)&&(i=e,e=""),i||(i=this[e]);var a=this;return t.history.route(n,function(r){var s=a._extractParameters(n,r);a.execute(i,s),a.trigger.apply(a,["route:"+e].concat(s)),a.trigger("route",e,s),t.history.trigger("route",a,e,s)}),this},execute:function(n,t){n&&n.apply(this,t)},navigate:function(n,r){return t.history.navigate(n,r),this},_bindRoutes:function(){if(this.routes){this.routes=r.result(this,"routes");for(var n,t=r.keys(this.routes);null!=(n=t.pop());)this.route(n,this.routes[n])}},_routeToRegExp:function(n){return n=n.replace(T,"\\$&").replace(A,"(?:$1)?").replace(S,function(n,t){return t?n:"([^/?]+)"}).replace(E,"([^?]*?)"),new RegExp("^"+n+"(?:\\?([\\s\\S]*))?$")},_extractParameters:function(n,t){var e=n.exec(t).slice(1);return r.map(e,function(n,t){return t===e.length-1?n||null:n?decodeURIComponent(n):null})}});var O=t.History=function(){this.handlers=[],r.bindAll(this,"checkUrl"),"undefined"!=typeof window&&(this.location=window.location,this.history=window.history)},$=/^[#\/]|\s+$/g,C=/^\/+|\/+$/g,F=/msie [\w.]+/,V=/\/$/,N=/#.*$/;O.started=!1,r.extend(O.prototype,l,{interval:50,atRoot:function(){return this.location.pathname.replace(/[^\/]$/,"$&/")===this.root},getHash:function(n){var t=(n||this).location.href.match(/#(.*)$/);return t?t[1]:""},getFragment:function(n,t){if(null==n)if(this._hasPushState||!this._wantsHashChange||t){n=decodeURI(this.location.pathname+this.location.search);var r=this.root.replace(V,"");n.indexOf(r)||(n=n.slice(r.length))}else n=this.getHash();return n.replace($,"")},start:function(n){if(O.started)throw new Error("Backbone.history has already been started");O.started=!0,this.options=r.extend({root:"/"},this.options,n),this.root=this.options.root,this._wantsHashChange=this.options.hashChange!==!1,this._wantsPushState=!!this.options.pushState,this._hasPushState=!!(this.options.pushState&&this.history&&this.history.pushState);var e=this.getFragment(),i=document.documentMode,a=F.exec(navigator.userAgent.toLowerCase())&&(!i||7>=i);if(this.root=("/"+this.root+"/").replace(C,"/"),a&&this._wantsHashChange){var s=t.$('<iframe src="javascript:0" tabindex="-1">');this.iframe=s.hide().appendTo("body")[0].contentWindow,this.navigate(e)}this._hasPushState?t.$(window).on("popstate",this.checkUrl):this._wantsHashChange&&"onhashchange"in window&&!a?t.$(window).on("hashchange",this.checkUrl):this._wantsHashChange&&(this._checkUrlInterval=setInterval(this.checkUrl,this.interval)),this.fragment=e;var l=this.location;if(this._wantsHashChange&&this._wantsPushState){if(!this._hasPushState&&!this.atRoot())return this.fragment=this.getFragment(null,!0),this.location.replace(this.root+"#"+this.fragment),!0;this._hasPushState&&this.atRoot()&&l.hash&&(this.fragment=this.getHash().replace($,""),this.history.replaceState({},document.title,this.root+this.fragment))}return this.options.silent?void 0:this.loadUrl()},stop:function(){t.$(window).off("popstate",this.checkUrl).off("hashchange",this.checkUrl),this._checkUrlInterval&&clearInterval(this._checkUrlInterval),O.started=!1},route:function(n,t){this.handlers.unshift({route:n,callback:t})},checkUrl:function(){var n=this.getFragment();return n===this.fragment&&this.iframe&&(n=this.getFragment(this.getHash(this.iframe))),n===this.fragment?!1:(this.iframe&&this.navigate(n),void this.loadUrl())},loadUrl:function(n){return n=this.fragment=this.getFragment(n),r.any(this.handlers,function(t){return t.route.test(n)?(t.callback(n),!0):void 0})},navigate:function(n,t){if(!O.started)return!1;t&&t!==!0||(t={trigger:!!t});var r=this.root+(n=this.getFragment(n||""));if(n=n.replace(N,""),this.fragment!==n){if(this.fragment=n,""===n&&"/"!==r&&(r=r.slice(0,-1)),this._hasPushState)this.history[t.replace?"replaceState":"pushState"]({},document.title,r);else{if(!this._wantsHashChange)return this.location.assign(r);this._updateHash(this.location,n,t.replace),this.iframe&&n!==this.getFragment(this.getHash(this.iframe))&&(t.replace||this.iframe.document.open().close(),this._updateHash(this.iframe.location,n,t.replace))}return t.trigger?this.loadUrl(n):void 0}},_updateHash:function(n,t,r){if(r){var e=n.href.replace(/(javascript:|#).*$/,"");n.replace(e+"#"+t)}else n.hash="#"+t}}),t.history=new O;var I=function(n,t){var e,i=this;e=n&&r.has(n,"constructor")?n.constructor:function(){return i.apply(this,arguments)},r.extend(e,i,t);var a=function(){this.constructor=e};return a.prototype=i.prototype,e.prototype=new a,n&&r.extend(e.prototype,n),e.__super__=i.prototype,e};f.extend=p.extend=j.extend=y.extend=O.extend=I;var M=function(){throw new Error('A "url" property or function must be specified')},J=function(n,t){var r=t.error;t.error=function(e){r&&r(n,e,t),n.trigger("error",n,e,t)}};return t})},{underscore:29}],29:[function(n,t,r){(function(){function n(n){function t(t,r,e,i,a,s){for(;a>=0&&s>a;a+=n){var l=i?i[a]:a;e=r(e,t[l],l,t)}return e}return function(r,e,i,a){e=w(e,a,4);var s=!A(r)&&y.keys(r),l=(s||r).length,o=n>0?0:l-1;return arguments.length<3&&(i=r[s?s[o]:o],o+=n),t(r,e,i,s,o,l)}}function e(n){return function(t,r,e){r=x(r,e);for(var i=null!=t&&t.length,a=n>0?0:i-1;a>=0&&i>a;a+=n)if(r(t[a],a,t))return a;return-1}}function i(n,t){var r=$.length,e=n.constructor,i=y.isFunction(e)&&e.prototype||o,a="constructor";for(y.has(n,a)&&!y.contains(t,a)&&t.push(a);r--;)a=$[r],a in n&&n[a]!==i[a]&&!y.contains(t,a)&&t.push(a)}var a=this,s=a._,l=Array.prototype,o=Object.prototype,c=Function.prototype,u=l.push,d=l.slice,f=o.toString,h=o.hasOwnProperty,p=Array.isArray,v=Object.keys,m=c.bind,g=Object.create,b=function(){},y=function(n){return n instanceof y?n:this instanceof y?void(this._wrapped=n):new y(n)};"undefined"!=typeof r?("undefined"!=typeof t&&t.exports&&(r=t.exports=y),r._=y):a._=y,y.VERSION="1.8.2";var w=function(n,t,r){if(void 0===t)return n;switch(null==r?3:r){case 1:return function(r){return n.call(t,r)};case 2:return function(r,e){return n.call(t,r,e)};case 3:return function(r,e,i){return n.call(t,r,e,i)};case 4:return function(r,e,i,a){return n.call(t,r,e,i,a)}}return function(){return n.apply(t,arguments)}},x=function(n,t,r){return null==n?y.identity:y.isFunction(n)?w(n,t,r):y.isObject(n)?y.matcher(n):y.property(n)};y.iteratee=function(n,t){return x(n,t,1/0)};var _=function(n,t){return function(r){var e=arguments.length;if(2>e||null==r)return r;for(var i=1;e>i;i++)for(var a=arguments[i],s=n(a),l=s.length,o=0;l>o;o++){var c=s[o];t&&void 0!==r[c]||(r[c]=a[c])}return r}},k=function(n){if(!y.isObject(n))return{};if(g)return g(n);b.prototype=n;var t=new b;return b.prototype=null,t},j=Math.pow(2,53)-1,A=function(n){var t=n&&n.length;return"number"==typeof t&&t>=0&&j>=t};y.each=y.forEach=function(n,t,r){t=w(t,r);var e,i;if(A(n))for(e=0,i=n.length;i>e;e++)t(n[e],e,n);else{var a=y.keys(n);for(e=0,i=a.length;i>e;e++)t(n[a[e]],a[e],n)}return n},y.map=y.collect=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=Array(i),s=0;i>s;s++){var l=e?e[s]:s;a[s]=t(n[l],l,n)}return a},y.reduce=y.foldl=y.inject=n(1),y.reduceRight=y.foldr=n(-1),y.find=y.detect=function(n,t,r){var e;return e=A(n)?y.findIndex(n,t,r):y.findKey(n,t,r),void 0!==e&&-1!==e?n[e]:void 0},y.filter=y.select=function(n,t,r){var e=[];return t=x(t,r),y.each(n,function(n,r,i){t(n,r,i)&&e.push(n)}),e},y.reject=function(n,t,r){return y.filter(n,y.negate(x(t)),r)},y.every=y.all=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=0;i>a;a++){var s=e?e[a]:a;if(!t(n[s],s,n))return!1}return!0},y.some=y.any=function(n,t,r){t=x(t,r);for(var e=!A(n)&&y.keys(n),i=(e||n).length,a=0;i>a;a++){var s=e?e[a]:a;if(t(n[s],s,n))return!0}return!1},y.contains=y.includes=y.include=function(n,t,r){return A(n)||(n=y.values(n)),y.indexOf(n,t,"number"==typeof r&&r)>=0},y.invoke=function(n,t){var r=d.call(arguments,2),e=y.isFunction(t);return y.map(n,function(n){var i=e?t:n[t];return null==i?i:i.apply(n,r)})},y.pluck=function(n,t){return y.map(n,y.property(t))},y.where=function(n,t){return y.filter(n,y.matcher(t))},y.findWhere=function(n,t){return y.find(n,y.matcher(t))},y.max=function(n,t,r){var e,i,a=-1/0,s=-1/0;if(null==t&&null!=n){n=A(n)?n:y.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],e>a&&(a=e)}else t=x(t,r),y.each(n,function(n,r,e){i=t(n,r,e),(i>s||i===-1/0&&a===-1/0)&&(a=n,s=i)});return a},y.min=function(n,t,r){var e,i,a=1/0,s=1/0;if(null==t&&null!=n){n=A(n)?n:y.values(n);for(var l=0,o=n.length;o>l;l++)e=n[l],a>e&&(a=e)}else t=x(t,r),y.each(n,function(n,r,e){i=t(n,r,e),(s>i||1/0===i&&1/0===a)&&(a=n,s=i)});return a},y.shuffle=function(n){for(var t,r=A(n)?n:y.values(n),e=r.length,i=Array(e),a=0;e>a;a++)t=y.random(0,a),t!==a&&(i[a]=i[t]),i[t]=r[a];return i},y.sample=function(n,t,r){return null==t||r?(A(n)||(n=y.values(n)),n[y.random(n.length-1)]):y.shuffle(n).slice(0,Math.max(0,t))},y.sortBy=function(n,t,r){return t=x(t,r),y.pluck(y.map(n,function(n,r,e){return{value:n,index:r,criteria:t(n,r,e)}}).sort(function(n,t){var r=n.criteria,e=t.criteria;if(r!==e){if(r>e||void 0===r)return 1;if(e>r||void 0===e)return-1}return n.index-t.index}),"value")};var S=function(n){return function(t,r,e){var i={};return r=x(r,e),y.each(t,function(e,a){var s=r(e,a,t);n(i,e,s)}),i}};y.groupBy=S(function(n,t,r){y.has(n,r)?n[r].push(t):n[r]=[t]}),y.indexBy=S(function(n,t,r){n[r]=t}),y.countBy=S(function(n,t,r){y.has(n,r)?n[r]++:n[r]=1}),y.toArray=function(n){return n?y.isArray(n)?d.call(n):A(n)?y.map(n,y.identity):y.values(n):[]},y.size=function(n){return null==n?0:A(n)?n.length:y.keys(n).length},y.partition=function(n,t,r){t=x(t,r);var e=[],i=[];return y.each(n,function(n,r,a){(t(n,r,a)?e:i).push(n)}),[e,i]},y.first=y.head=y.take=function(n,t,r){return null==n?void 0:null==t||r?n[0]:y.initial(n,n.length-t)},y.initial=function(n,t,r){return d.call(n,0,Math.max(0,n.length-(null==t||r?1:t)))},y.last=function(n,t,r){return null==n?void 0:null==t||r?n[n.length-1]:y.rest(n,Math.max(0,n.length-t))},y.rest=y.tail=y.drop=function(n,t,r){return d.call(n,null==t||r?1:t)},y.compact=function(n){return y.filter(n,y.identity)};var E=function(n,t,r,e){for(var i=[],a=0,s=e||0,l=n&&n.length;l>s;s++){var o=n[s];if(A(o)&&(y.isArray(o)||y.isArguments(o))){t||(o=E(o,t,r));var c=0,u=o.length;for(i.length+=u;u>c;)i[a++]=o[c++]}else r||(i[a++]=o)}return i};y.flatten=function(n,t){return E(n,t,!1)},y.without=function(n){return y.difference(n,d.call(arguments,1))},y.uniq=y.unique=function(n,t,r,e){if(null==n)return[];y.isBoolean(t)||(e=r,r=t,t=!1),null!=r&&(r=x(r,e));for(var i=[],a=[],s=0,l=n.length;l>s;s++){var o=n[s],c=r?r(o,s,n):o;t?(s&&a===c||i.push(o),a=c):r?y.contains(a,c)||(a.push(c),i.push(o)):y.contains(i,o)||i.push(o)}return i},y.union=function(){return y.uniq(E(arguments,!0,!0))},y.intersection=function(n){if(null==n)return[];for(var t=[],r=arguments.length,e=0,i=n.length;i>e;e++){var a=n[e];if(!y.contains(t,a)){for(var s=1;r>s&&y.contains(arguments[s],a);s++);s===r&&t.push(a)}}return t},y.difference=function(n){var t=E(arguments,!0,!0,1);return y.filter(n,function(n){return!y.contains(t,n)})},y.zip=function(){return y.unzip(arguments)},y.unzip=function(n){for(var t=n&&y.max(n,"length").length||0,r=Array(t),e=0;t>e;e++)r[e]=y.pluck(n,e);return r},y.object=function(n,t){for(var r={},e=0,i=n&&n.length;i>e;e++)t?r[n[e]]=t[e]:r[n[e][0]]=n[e][1];return r},y.indexOf=function(n,t,r){var e=0,i=n&&n.length;if("number"==typeof r)e=0>r?Math.max(0,i+r):r;else if(r&&i)return e=y.sortedIndex(n,t),n[e]===t?e:-1;if(t!==t)return y.findIndex(d.call(n,e),y.isNaN);for(;i>e;e++)if(n[e]===t)return e;return-1},y.lastIndexOf=function(n,t,r){var e=n?n.length:0;if("number"==typeof r&&(e=0>r?e+r+1:Math.min(e,r+1)),t!==t)return y.findLastIndex(d.call(n,0,e),y.isNaN);for(;--e>=0;)if(n[e]===t)return e;return-1},y.findIndex=e(1),y.findLastIndex=e(-1),y.sortedIndex=function(n,t,r,e){r=x(r,e,1);for(var i=r(t),a=0,s=n.length;s>a;){var l=Math.floor((a+s)/2);r(n[l])<i?a=l+1:s=l}return a},y.range=function(n,t,r){arguments.length<=1&&(t=n||0,n=0),r=r||1;for(var e=Math.max(Math.ceil((t-n)/r),0),i=Array(e),a=0;e>a;a++,n+=r)i[a]=n;return i};var T=function(n,t,r,e,i){if(!(e instanceof t))return n.apply(r,i);var a=k(n.prototype),s=n.apply(a,i);return y.isObject(s)?s:a};y.bind=function(n,t){if(m&&n.bind===m)return m.apply(n,d.call(arguments,1));if(!y.isFunction(n))throw new TypeError("Bind must be called on a function");var r=d.call(arguments,2),e=function(){return T(n,e,t,this,r.concat(d.call(arguments)))};return e},y.partial=function(n){var t=d.call(arguments,1),r=function(){for(var e=0,i=t.length,a=Array(i),s=0;i>s;s++)a[s]=t[s]===y?arguments[e++]:t[s];for(;e<arguments.length;)a.push(arguments[e++]);return T(n,r,this,this,a)};return r},y.bindAll=function(n){var t,r,e=arguments.length;if(1>=e)throw new Error("bindAll must be passed function names");for(t=1;e>t;t++)r=arguments[t],n[r]=y.bind(n[r],n);return n},y.memoize=function(n,t){var r=function(e){var i=r.cache,a=""+(t?t.apply(this,arguments):e);return y.has(i,a)||(i[a]=n.apply(this,arguments)),i[a]};return r.cache={},r},y.delay=function(n,t){var r=d.call(arguments,2);return setTimeout(function(){return n.apply(null,r)},t)},y.defer=y.partial(y.delay,y,1),y.throttle=function(n,t,r){var e,i,a,s=null,l=0;r||(r={});var o=function(){l=r.leading===!1?0:y.now(),s=null,a=n.apply(e,i),s||(e=i=null)};return function(){var c=y.now();l||r.leading!==!1||(l=c);var u=t-(c-l);return e=this,i=arguments,0>=u||u>t?(s&&(clearTimeout(s),s=null),l=c,a=n.apply(e,i),s||(e=i=null)):s||r.trailing===!1||(s=setTimeout(o,u)),a}},y.debounce=function(n,t,r){var e,i,a,s,l,o=function(){var c=y.now()-s;t>c&&c>=0?e=setTimeout(o,t-c):(e=null,r||(l=n.apply(a,i),e||(a=i=null)))};return function(){a=this,i=arguments,s=y.now();var c=r&&!e;return e||(e=setTimeout(o,t)),c&&(l=n.apply(a,i),a=i=null),l}},y.wrap=function(n,t){return y.partial(t,n)},y.negate=function(n){return function(){return!n.apply(this,arguments)}},y.compose=function(){var n=arguments,t=n.length-1;return function(){for(var r=t,e=n[t].apply(this,arguments);r--;)e=n[r].call(this,e);return e}},y.after=function(n,t){return function(){return--n<1?t.apply(this,arguments):void 0}},y.before=function(n,t){var r;return function(){return--n>0&&(r=t.apply(this,arguments)),1>=n&&(t=null),r}},y.once=y.partial(y.before,2);var O=!{toString:null}.propertyIsEnumerable("toString"),$=["valueOf","isPrototypeOf","toString","propertyIsEnumerable","hasOwnProperty","toLocaleString"];y.keys=function(n){if(!y.isObject(n))return[];if(v)return v(n);var t=[];for(var r in n)y.has(n,r)&&t.push(r);return O&&i(n,t),t},y.allKeys=function(n){if(!y.isObject(n))return[];var t=[];for(var r in n)t.push(r);return O&&i(n,t),t},y.values=function(n){for(var t=y.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=n[t[i]];return e},y.mapObject=function(n,t,r){t=x(t,r);for(var e,i=y.keys(n),a=i.length,s={},l=0;a>l;l++)e=i[l],s[e]=t(n[e],e,n);return s},y.pairs=function(n){for(var t=y.keys(n),r=t.length,e=Array(r),i=0;r>i;i++)e[i]=[t[i],n[t[i]]];return e},y.invert=function(n){for(var t={},r=y.keys(n),e=0,i=r.length;i>e;e++)t[n[r[e]]]=r[e];return t},y.functions=y.methods=function(n){var t=[];for(var r in n)y.isFunction(n[r])&&t.push(r);return t.sort()},y.extend=_(y.allKeys),y.extendOwn=y.assign=_(y.keys),y.findKey=function(n,t,r){t=x(t,r);for(var e,i=y.keys(n),a=0,s=i.length;s>a;a++)if(e=i[a],t(n[e],e,n))return e},y.pick=function(n,t,r){var e,i,a={},s=n;if(null==s)return a;y.isFunction(t)?(i=y.allKeys(s),e=w(t,r)):(i=E(arguments,!1,!1,1),e=function(n,t,r){return t in r},s=Object(s));for(var l=0,o=i.length;o>l;l++){var c=i[l],u=s[c];e(u,c,s)&&(a[c]=u)}return a},y.omit=function(n,t,r){if(y.isFunction(t))t=y.negate(t);else{var e=y.map(E(arguments,!1,!1,1),String);t=function(n,t){return!y.contains(e,t)}}return y.pick(n,t,r)},y.defaults=_(y.allKeys,!0),y.clone=function(n){return y.isObject(n)?y.isArray(n)?n.slice():y.extend({},n):n},y.tap=function(n,t){return t(n),n},y.isMatch=function(n,t){var r=y.keys(t),e=r.length;if(null==n)return!e;for(var i=Object(n),a=0;e>a;a++){var s=r[a];if(t[s]!==i[s]||!(s in i))return!1}return!0};var C=function(n,t,r,e){if(n===t)return 0!==n||1/n===1/t;if(null==n||null==t)return n===t;n instanceof y&&(n=n._wrapped),t instanceof y&&(t=t._wrapped);var i=f.call(n);if(i!==f.call(t))return!1;switch(i){case"[object RegExp]":case"[object String]":return""+n==""+t;case"[object Number]":return+n!==+n?+t!==+t:0===+n?1/+n===1/t:+n===+t;case"[object Date]":case"[object Boolean]":return+n===+t}var a="[object Array]"===i;if(!a){if("object"!=typeof n||"object"!=typeof t)return!1;var s=n.constructor,l=t.constructor;if(s!==l&&!(y.isFunction(s)&&s instanceof s&&y.isFunction(l)&&l instanceof l)&&"constructor"in n&&"constructor"in t)return!1}r=r||[],e=e||[];for(var o=r.length;o--;)if(r[o]===n)return e[o]===t;if(r.push(n),e.push(t),a){if(o=n.length,o!==t.length)return!1;for(;o--;)if(!C(n[o],t[o],r,e))return!1}else{var c,u=y.keys(n);if(o=u.length,y.keys(t).length!==o)return!1;for(;o--;)if(c=u[o],!y.has(t,c)||!C(n[c],t[c],r,e))return!1}return r.pop(),e.pop(),!0};y.isEqual=function(n,t){return C(n,t)},y.isEmpty=function(n){return null==n?!0:A(n)&&(y.isArray(n)||y.isString(n)||y.isArguments(n))?0===n.length:0===y.keys(n).length},y.isElement=function(n){return!(!n||1!==n.nodeType)},y.isArray=p||function(n){return"[object Array]"===f.call(n)
+},y.isObject=function(n){var t=typeof n;return"function"===t||"object"===t&&!!n},y.each(["Arguments","Function","String","Number","Date","RegExp","Error"],function(n){y["is"+n]=function(t){return f.call(t)==="[object "+n+"]"}}),y.isArguments(arguments)||(y.isArguments=function(n){return y.has(n,"callee")}),"function"!=typeof/./&&"object"!=typeof Int8Array&&(y.isFunction=function(n){return"function"==typeof n||!1}),y.isFinite=function(n){return isFinite(n)&&!isNaN(parseFloat(n))},y.isNaN=function(n){return y.isNumber(n)&&n!==+n},y.isBoolean=function(n){return n===!0||n===!1||"[object Boolean]"===f.call(n)},y.isNull=function(n){return null===n},y.isUndefined=function(n){return void 0===n},y.has=function(n,t){return null!=n&&h.call(n,t)},y.noConflict=function(){return a._=s,this},y.identity=function(n){return n},y.constant=function(n){return function(){return n}},y.noop=function(){},y.property=function(n){return function(t){return null==t?void 0:t[n]}},y.propertyOf=function(n){return null==n?function(){}:function(t){return n[t]}},y.matcher=y.matches=function(n){return n=y.extendOwn({},n),function(t){return y.isMatch(t,n)}},y.times=function(n,t,r){var e=Array(Math.max(0,n));t=w(t,r,1);for(var i=0;n>i;i++)e[i]=t(i);return e},y.random=function(n,t){return null==t&&(t=n,n=0),n+Math.floor(Math.random()*(t-n+1))},y.now=Date.now||function(){return(new Date).getTime()};var F={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;","'":"&#x27;","`":"&#x60;"},V=y.invert(F),N=function(n){var t=function(t){return n[t]},r="(?:"+y.keys(n).join("|")+")",e=RegExp(r),i=RegExp(r,"g");return function(n){return n=null==n?"":""+n,e.test(n)?n.replace(i,t):n}};y.escape=N(F),y.unescape=N(V),y.result=function(n,t,r){var e=null==n?void 0:n[t];return void 0===e&&(e=r),y.isFunction(e)?e.call(n):e};var I=0;y.uniqueId=function(n){var t=++I+"";return n?n+t:t},y.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g,escape:/<%-([\s\S]+?)%>/g};var M=/(.)^/,J={"'":"'","\\":"\\","\r":"r","\n":"n","\u2028":"u2028","\u2029":"u2029"},q=/\\|'|\r|\n|\u2028|\u2029/g,R=function(n){return"\\"+J[n]};y.template=function(n,t,r){!t&&r&&(t=r),t=y.defaults({},t,y.templateSettings);var e=RegExp([(t.escape||M).source,(t.interpolate||M).source,(t.evaluate||M).source].join("|")+"|$","g"),i=0,a="__p+='";n.replace(e,function(t,r,e,s,l){return a+=n.slice(i,l).replace(q,R),i=l+t.length,r?a+="'+\n((__t=("+r+"))==null?'':_.escape(__t))+\n'":e?a+="'+\n((__t=("+e+"))==null?'':__t)+\n'":s&&(a+="';\n"+s+"\n__p+='"),t}),a+="';\n",t.variable||(a="with(obj||{}){\n"+a+"}\n"),a="var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n"+a+"return __p;\n";try{var s=new Function(t.variable||"obj","_",a)}catch(l){throw l.source=a,l}var o=function(n){return s.call(this,n,y)},c=t.variable||"obj";return o.source="function("+c+"){\n"+a+"}",o},y.chain=function(n){var t=y(n);return t._chain=!0,t};var P=function(n,t){return n._chain?y(t).chain():t};y.mixin=function(n){y.each(y.functions(n),function(t){var r=y[t]=n[t];y.prototype[t]=function(){var n=[this._wrapped];return u.apply(n,arguments),P(this,r.apply(y,n))}})},y.mixin(y),y.each(["pop","push","reverse","shift","sort","splice","unshift"],function(n){var t=l[n];y.prototype[n]=function(){var r=this._wrapped;return t.apply(r,arguments),"shift"!==n&&"splice"!==n||0!==r.length||delete r[0],P(this,r)}}),y.each(["concat","join","slice"],function(n){var t=l[n];y.prototype[n]=function(){return P(this,t.apply(this._wrapped,arguments))}}),y.prototype.value=function(){return this._wrapped},y.prototype.valueOf=y.prototype.toJSON=y.prototype.value,y.prototype.toString=function(){return""+this._wrapped},"function"==typeof define&&define.amd&&define("underscore",[],function(){return y})}).call(this)},{}],30:[function(n,t,r){arguments[4][29][0].apply(r,arguments)},{dup:29}]},{},[1]);
 //# sourceMappingURL=app.js.map
